% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcDailySpawnProps2_SpDurMod}
\alias{CalcDailySpawnProps2_SpDurMod}
\title{Calculate daily proportions of spawning capable fish of a given length, using spawning duration model,
used for calculating NLL}
\usage{
CalcDailySpawnProps2_SpDurMod(params)
}
\arguments{
\item{params}{specified (or estimated), transformed parameter values, for the spawning duration model (including
lnL50, lnslope, logitPkSpawn, lnkappa, lnkappa2, lnslope1, lnslope2)}
}
\value{
returns a list object containing expected proportions of spawning capable fish, of a given length,
for each specified day of the year, inputted as decimal day (P_t_s), and the value of a penalty
function used to ensure the maximum proportion of spawning capable fishing is always between 0 and 1
}
\description{
This function returns calculated values for the daily proportions of spawning capable fish
of a specified length, using spawning duration model (i.e. asymmetric double logistic model with
variable height)
}
\keyword{internal}
