% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcDailySpawnProps_SpDurMod}
\alias{CalcDailySpawnProps_SpDurMod}
\title{Calculate daily proportions of spawning capable fish of a given length, using spawning duration model}
\usage{
CalcDailySpawnProps_SpDurMod(fish_size, DecDay, params)
}
\arguments{
\item{params}{specified (or estimated), transformed parameter values, for the spawning duration model (including
lnL50, lnslope, logitPkSpawn, lnkappa, lnkappa2, lnslope1, lnslope2)}

\item{FishLen}{specified fish length}

\item{DistnType}{DecDay specified decimal days of the year}
}
\value{
returns a list object containing expected proportions of spawning capable fish, of a given length,
for each specified day of the year, inputted as decimal day (P_t_s), and the value of a penalty
function used to ensure the maximum proportion of spawning capable fishing is always between 0 and 1
}
\description{
This function returns calculated values for the daily proportions of spawning capable fish
of a specified length, using spawning duration model (i.e. asymmetric double logistic model with
variable height)
}
\examples{
FishLen = 500
DecDay = seq(0,1,0.01)
lnL50 = log(300)
lnslope= log(0.015)
logitPkSpawn = log(0.6/(1-0.6))
lnkappa = log(0.0002)
lnkappa2 = log(0.0002)
lnslope1 = log(30)
lnslope2 = log(30)
params = c(lnL50,lnslope,logitPkSpawn,lnkappa,lnkappa2,lnslope1,lnslope2)
res=CalcDailySpawnProps_SpDurMod(FishLen, DecDay, params)
plot(DecDay,res$P_t_s,"l")
}
