% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcLengthAtAge_DivergeGrowthModel}
\alias{CalcLengthAtAge_DivergeGrowthModel}
\title{Calculate negative log-likelihood associated with fit of growth model allowing for divergent growth between sexes or areas}
\usage{
CalcLengthAtAge_DivergeGrowthModel(params)
}
\arguments{
\item{params}{c(log(c(Linf,Linf)),log(c(vbK,vbK)),tzero1,log(tdiverge))}
}
\value{
Negative log-likelihood
}
\description{
Calculates the negative log-likelihood associated with a sample of fish length-at-age data
and associated growth curve parameter values, for a growth model allowing for divergent
growth, after a certain age, between sexes or fish in different areas
}
\keyword{internal}
