% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcLengthAtAge_DivergentGrowthModel2}
\alias{CalcLengthAtAge_DivergentGrowthModel2}
\title{Calculate expected lengths at age from a growth model allowing for divergent growth}
\usage{
CalcLengthAtAge_DivergentGrowthModel2(params, plotages)
}
\arguments{
\item{params}{c(log(Linf),log(vbK),tzero, log(tdiverge))}

\item{plotages}{specified ages for plotting}
}
\value{
specified ages (plotages) and expected lengths at ages (plotlengths)
}
\description{
Calculates expected lengths at age from von Bertalanffy growth model
allowing for divergent growth between sexes (or areas), used for plotting
}
\keyword{internal}
