% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcLengthAtAge_SchnuteGrowthCurve}
\alias{CalcLengthAtAge_SchnuteGrowthCurve}
\title{Calculate expected lengths at age from Schnute growth curve}
\usage{
CalcLengthAtAge_SchnuteGrowthCurve(params, t1, t2, ObsAge)
}
\arguments{
\item{params}{c(log(y1),log(y2),a,b) Schnute parameters}

\item{t1}{reference age 1}

\item{t2}{reference age 2}

\item{ObsAge}{specified ages}
}
\value{
expected lengths at age (ExpLen)
}
\description{
Calculates expected lengths at age from Schnute growth curve
and associated growth curve parameter values. Used to calculated expected lengths
for observed ages
}
\keyword{internal}
