% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcLengthAtAge_SchnuteGrowthCurve2}
\alias{CalcLengthAtAge_SchnuteGrowthCurve2}
\title{Calculate expected lengths at age from Schnute growth curve}
\usage{
CalcLengthAtAge_SchnuteGrowthCurve2(params, nSexes, Ref_ages, plotages)
}
\arguments{
\item{params}{c(log(y1),log(y2),a,b) Schnute parameters}

\item{nSexes}{number of sexes}

\item{Ref_ages}{Schnute growth curve reference ages (t1 and t2)}

\item{plotages}{specified ages for plotting}

\item{ObsAge}{specified ages}
}
\value{
specified ages (plotages) and expected lengths at ages (plotlengths)
}
\description{
Calculates expected lengths at age from Schnute growth curve
and associated growth curve parameter values. Used to calculated expected lengths
at specified ages for plotting.
}
\keyword{internal}
