% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcLengthAtAge_SomersSeasonalGrowthCurve}
\alias{CalcLengthAtAge_SomersSeasonalGrowthCurve}
\title{Get expected lengths at age from a seasonal growth curve}
\usage{
CalcLengthAtAge_SomersSeasonalGrowthCurve(params)
}
\arguments{
\item{params}{c(log(Linf), log(vbK), tzero, tc, C)}
}
\value{
Returns expected lengths at age from Somers (1988) seasonal growth curve
}
\description{
Returns expected lengths at age from Somers (1988) seasonal growth curve, given parameter values
and ages
}
\examples{
# Generate length at age data for individual fish
DataType = 1
set.seed(123)
nSamples = 300
ObsAge = runif(nSamples, 0.1, 3)
params = c(log(150),log(1),0,0.25,0.8) #Linf, vbK, tzero, tc, C
MeanLen = CalcLengthAtAge_SomersSeasonalGrowthCurve(params)
ObsLen = rnorm(nSamples, MeanLen, 0.1*MeanLen)
plot(ObsAge, ObsLen)
params = c(log(150),log(1),0,0.25,0.8) #Linf, vbK, tzero, tc, C
res=CalcLengthAtAge_SomersSeasonalGrowthCurve(params)
points(ObsAge, res)
# Generate monthly mean length (with error) data, from mixture analysis
DataType = 2
set.seed(123)
ObsAge = seq(0.2, 3, 1/12)
nAges=length(ObsAge)
params = c(log(150),log(1),0,0.25,0.8) #Linf, vbK, tzero, tc, C
ObsMeanLen = CalcLengthAtAge_SomersSeasonalGrowthCurve(params)
ObsMeanLensd = rnorm(nAges,5,1)
ObsLenlw = ObsMeanLen - (1.96 * ObsMeanLensd)
ObsLenup = ObsMeanLen + (1.96 * ObsMeanLensd)
ymax=1.1*max(ObsLenup)
plot(ObsAge, ObsMeanLen, ylim=c(0,ymax))
arrows(ObsAge, ObsLenlw, ObsAge, ObsLenup,
       code=3, angle=90,length=0.02, col='black')
params = c(log(150),log(1),0,0.25,0.8) #Linf, vbK, tzero, tc, C
res=CalcLengthAtAge_SomersSeasonalGrowthCurve(params)
lines(ObsAge,res)
}
