% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcLengthAtAge_SomersSeasonalGrowthCurve2}
\alias{CalcLengthAtAge_SomersSeasonalGrowthCurve2}
\title{Calculate expected lengths at age from seasonal growth curve}
\usage{
CalcLengthAtAge_SomersSeasonalGrowthCurve2(params, nSexes, plotages)
}
\arguments{
\item{params}{number of sexes}

\item{plotages}{specified ages for plotting}
}
\value{
specified ages (plotages) and expected lengths at age (plotlengths)
}
\description{
Calculates expected lengths at age from Somers seasonal growth curve
and associated growth curve parameter values. Used in plotting where
ages are pres-specified
}
\keyword{internal}
