% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcLengthAtAge_vonBertalanffyGrowthCurve}
\alias{CalcLengthAtAge_vonBertalanffyGrowthCurve}
\title{Calculate expected lengths at age from von Bertalanffy growth curve}
\usage{
CalcLengthAtAge_vonBertalanffyGrowthCurve(params)
}
\arguments{
\item{params}{c(log(Linf),log(vbK),tzero) single or combined sex, or c(log(c(Linf,Linf)),log(c(vbK,vbK)),c(tzero,tzero))}
}
\value{
expected lengths at age (ExpLen)
}
\description{
Calculates expected lengths at age from von Bertalanffy growth curve
and associated growth curve parameter values. Used to calculated expected lengths
for observed ages
}
\keyword{internal}
