% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcLengthAtAge_vonBertalanffyGrowthCurve2}
\alias{CalcLengthAtAge_vonBertalanffyGrowthCurve2}
\title{Calculate expected lengths at age from von Bertalanffy growth curve}
\usage{
CalcLengthAtAge_vonBertalanffyGrowthCurve2(params, nSexes, plotages)
}
\arguments{
\item{params}{c(log(Linf),log(vbK),tzero)}

\item{nSexes}{number of sexes}

\item{plotages}{specified ages for plotting}
}
\value{
specified ages (plotages) and expected lengths at ages (plotlengths)
}
\description{
Calculates expected lengths at age from von Bertalanffy growth curve
and associated growth curve parameter values. Used in plotting where
ages are pres-specified
}
\keyword{internal}
