% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcMeanLenOff_OffMoveMod}
\alias{CalcMeanLenOff_OffMoveMod}
\title{Calculate expected mean length of fish in the adult habitat, given overall mean length
across habitats and specified observed length}
\usage{
CalcMeanLenOff_OffMoveMod(ObsLen, MeanLen, Growthsd, moveL50, moveL95)
}
\arguments{
\item{params}{x, MeanLen, Growthsd, moveL50, moveL95}
}
\value{
MeanOff
}
\description{
This function calculates the expected mean length of fish in the adult habitat, given overall mean length
across habitats and specified observed length
}
\keyword{internal}
