% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcMeanMonthlyGSIs}
\alias{CalcMeanMonthlyGSIs}
\title{Calculate mean monthly GSIs}
\usage{
CalcMeanMonthlyGSIs(MatL50, FishRepdat)
}
\arguments{
\item{MatL50}{length at first maturity}

\item{FishRepdat}{data frame containing data for individual fish, including month of capture (MM), fish length (FishLen),
gonad wet weight (GonadWt) and fish total body weight (FishWt)}
}
\value{
data frame (Result) containing elements month (MM), mean GSI (Mean), standard
deviation (sd), monthly sample size (n), standard error (se), lower and upper 95 percent
confidence limits (Low95 and Up95)
}
\description{
This function calculates mean monthly gonaosomatic indices GSI values and associated 95 percent confidence limits
}
\examples{
set.seed(123)
nMnths = 12
MnthSampeSizes = 20
MMSampSize = rep(MnthSampeSizes,nMnths)
TotSampSize = MnthSampeSizes * nMnths
Mnths = 1:nMnths
MeanMMGSI = rep(c(rep(0.1,4),0.25, 0.7, 0.6, 0.3, rep(0.1,4)),MMSampSize)
randMMGSI = exp(rnorm(TotSampSize,log(MeanMMGSI),0.5))
length(randMMGSI)
MM = rep(Mnths,times=MMSampSize)
FishLen = round(runif(TotSampSize,100,500),0)
FishWt = (0.0000001 * FishLen ^ 3)
GonadWt = (randMMGSI / 100) * FishWt
FishRepdat = data.frame(MM=MM, FishLen=FishLen, GonadWt=GonadWt, FishWt=FishWt)
MatL50 = 200
CalcMeanMonthlyGSIs(MatL50, FishRepdat)
}
