% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcMonthlyGonadStageProps}
\alias{CalcMonthlyGonadStageProps}
\title{Calculate monthly proportions of each gonad stage}
\usage{
CalcMonthlyGonadStageProps(MatL50, FishRepdat)
}
\arguments{
\item{MatL50}{length at first maturity}

\item{FishRepdat}{data frame containing data for individual fish, including month of capture (MM), fish length (FishLen),}
}
\value{
Monthly gonad stage proportions (PropGonadSt) with a list of gonad stages in data (GonadStages)
}
\description{
Calculates monthly proportions of each gonad stage, used to assess time and duration of spawning
}
