% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcMonthlyObsSpawnProps_SpDurMod}
\alias{CalcMonthlyObsSpawnProps_SpDurMod}
\title{Calculate observed monthly proportions of spawning capable fish}
\usage{
CalcMonthlyObsSpawnProps_SpDurMod(subDat)
}
\arguments{
\item{subDat}{data frame with observed months (MM) and maturity status (ObsMatStatus,
0=immature, 1=mature)}
}
\value{
observed monthly proportions of spawning capable fish (Probs)
}
\description{
This function calculates monthly proportions of spawning capable fish for observed
input data on months and observed maturity status
}
\keyword{internal}
