% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcNLL_DivergeGrowthModel}
\alias{CalcNLL_DivergeGrowthModel}
\title{Calculate negative log-likelihood associated with fit of a growth model allowing for divergent growth}
\usage{
CalcNLL_DivergeGrowthModel(params)
}
\arguments{
\item{params}{c(log(c(Linf,Linf)),log(c(vbK,vbK)),tzero,log(tdiverge))}
}
\value{
Negative log-likelihood associated with growth curve fit to length-at-age data
}
\description{
Calculates the negative log-likelihood associated with a sample of fish length-at-age data
and associated growth curve parameter values, for a growth models allowing for divergent
growth between sexes (or areas)
}
\keyword{internal}
