% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcNLL_GrowthCurve}
\alias{CalcNLL_GrowthCurve}
\title{Calculate negative log-likelihood associated with fit of a growth curve}
\usage{
CalcNLL_GrowthCurve(params)
}
\arguments{
\item{params}{c(log(Linf),log(vbK),tzero) single or combined sex, or c(log(c(Linf,Linf)),log(c(vbK,vbK)),c(tzero,tzero))}
}
\value{
Negative log-likelihood associated with growth curve fit to length-at-age data
}
\description{
Calculates the negative log-likelihood associated with a sample of fish length-at-age data
and associated growth curve parameter values. Growth models include von Bertalanffy,
Schnute and Somers seasonal growth curve.
}
\keyword{internal}
