% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcNLL_LogisticAgeAtMaturity}
\alias{CalcNLL_LogisticAgeAtMaturity}
\title{Calculate negative log-likelihood for an age-based logistic maturity curve.}
\usage{
CalcNLL_LogisticAgeAtMaturity(params)
}
\arguments{
\item{params}{c(A50, A95) ages at 50 and 95 percent maturity}
}
\value{
Negative-log likelihood associated with logistic curve fit to maturity-at-age data
}
\description{
Calculates the negative log-likelihood associated with a sample of fish maturity-at-age data
and logistic maturity parameter values (or sex change data).
}
\keyword{internal}
