% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcNLL_LogisticLengthAtMaturity}
\alias{CalcNLL_LogisticLengthAtMaturity}
\title{Calculate negative log-likelihood for a length-based logistic maturity curve.}
\usage{
CalcNLL_LogisticLengthAtMaturity(params)
}
\arguments{
\item{params}{c(L50, L95) lengths at 50 and 95 percent maturity}
}
\value{
Negative-log likelihood associated with logistic curve fit to maturity-at-length data
}
\description{
Calculates the negative log-likelihood associated with a sample of fish maturity-at-length data
and logistic maturity parameter values.
}
\keyword{internal}
