% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcNLL_OffMoveMod}
\alias{CalcNLL_OffMoveMod}
\title{Calculate negative log-likelihood associated with offshore-movement growth model}
\usage{
CalcNLL_OffMoveMod(params)
}
\arguments{
\item{params}{parameters to be estimated}
}
\value{
NLL, REPORT and ADREPORT variables
}
\description{
Calculate negative log-likelihood associated with offshore-movement growth model,
for use in RTMB.
}
\keyword{internal}
