% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcNLL_SizeMixtureDistn}
\alias{CalcNLL_SizeMixtureDistn}
\title{Calculate negative log-likelihood associated with the fit of a size mixture distribution model
and associated parameters values to size frequency data}
\usage{
CalcNLL_SizeMixtureDistn(params)
}
\arguments{
\item{params}{c(Mean1, sd1) normal distn: 1 mode, or
c(Shape1, Rate1) gamma distn: 1 mode, or
c(Mean1, Mean2, sd1, PropZero) normal distn: 2, common sd, or
c(Shape1, Shape2, Rate1, PropZero) gamma distn: 2, common sd, or
c(Mean1, Mean2, sd1, sd2, PropZero) normal distn: 2, separate sds, or
c(Shape1, Shape2, Rate1, Rate2, PropZero) gamma distn: 2 modes, separate sds, or
c(Mean1, Mean2, Mean3, sd1, sd2, sd3, PropZero, PropOne) normal distn: 3 modes, separate sds, or
c(Shape1, Shape2, Shape3, Rate1, Rate2, Rate3, PropZero, PropOne) gamma distn: 3 modes, separate sds}
}
\value{
Negative log-likelihood associated with mixture model fit to size frequency data
}
\description{
Calculates the negative log-likelihood associated with the fit of a size mixture
distribution model and associated parameters values to size frequency data,
with 1, 2 or 3 size modes. The model assumes each size mode is normally-distributed.
Four alternative models are available depending on the number of modes present in the
data, and assumptions regarding the spread of the various size modes.
Function requires size frequency data, and values for minimum size, maximum size
and size interval (stored in memory in R).
ObsFreq (Numeric Vector), MinSize, MaxSize, SizeInt (Numbers).
This function (with parameter inputs) can be passed into R optimisation routines (e.g. nlminb).
}
\keyword{internal}
