% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcNLL_SpDurMod}
\alias{CalcNLL_SpDurMod}
\title{Calculate NLL for spawning duration model given data and parameter values}
\usage{
CalcNLL_SpDurMod(params)
}
\arguments{
\item{params}{specified (or estimated), transformed parameter values, for the spawning duration model (including
lnL50, lnslope, logitPkSpawn, lnkappa, lnkappa2, lnslope1, lnslope2)}
}
\value{
negative log-likelihood (NLL)
}
\description{
Calculate negative log-likelihood for spawning duration model given data and parameter values
}
\keyword{internal}
