% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcNLL_TaggingGrowthModel}
\alias{CalcNLL_TaggingGrowthModel}
\title{Calculate negative log-likelihood for growth model fitted to tag-recapture data.}
\usage{
CalcNLL_TaggingGrowthModel(params)
}
\arguments{
\item{Model}{1: L50_1, L95_1, L50_2, L95_2, Max_increment (double logistic model)
Model 3: vb_Linf, vb_K (von Bertalanffy)
Model 4: Gomp_Linf, Gomp_G}
}
\value{
Negative-log likelihood associated with growth curve fit to tag-recapture data
}
\description{
Calculates the negative log-likelihood associated with a sample of animal tag-recapture data
and associated parameters of the growth model. Four alternative growth models are currently
available, i.e. von Bertalanffy curve, Gompertz curve, Gaussian function curve, Double logistic
function curve. Each describe the expected length increment between initial capture, measurement
and marking, and final capture and measurement.
Function requires data for individual animals on size at initial capture, size at final capture,
and time at liberty. Obs_Initlen, Obs_Finlen, Obs_delta_t (Numeric Vectors)
(stored in memory in R).
This function (with parameter inputs) can be passed into R optimisation routines (e.g. nlminb).
}
