% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcProbInsh_OffMoveMod}
\alias{CalcProbInsh_OffMoveMod}
\title{Calculate the probability of a fish being in a juvenile habitat from its length}
\usage{
CalcProbInsh_OffMoveMod(ObsLen, MeanLen, Growthsd, moveL50, moveL95)
}
\arguments{
\item{params}{x, MeanLen, Growthsd, moveL50, moveL95}
}
\value{
ProbInsh
}
\description{
This function calculates the probability of a fish being in a juvenile habitat from its length
}
\keyword{internal}
