% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcProbOff_OffMoveMod}
\alias{CalcProbOff_OffMoveMod}
\title{Calculate the probability of the expected length of a fish being in an adult habitat from its length}
\usage{
CalcProbOff_OffMoveMod(ObsLen, MeanLen, Growthsd, moveL50, moveL95)
}
\arguments{
\item{params}{x, MeanLen, Growthsd, moveL50, moveL95}
}
\value{
ProbOff
}
\description{
This function calculates the probability of a fish being in an adult habitat from its length
}
\keyword{internal}
