% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcPropMatureAtAge}
\alias{CalcPropMatureAtAge}
\title{Calculate proportions of fish mature in successive age classes}
\usage{
CalcPropMatureAtAge(MinAge, MaxAge, ObsAgeCl, ObsMatCat)
}
\arguments{
\item{MinAge}{minimum age class}

\item{MaxAge}{maximum age class}

\item{ObsAgeCl}{vector of observed age classes}

\item{ObsMatCat}{vector of observed maturity categories (0=immature, 1=mature)}
}
\value{
Age classes (AgeClasses), proportions of mature fish in successive age classes
(PropMat), ages for plotting maturity curve (plotages), age class sample sizes (AgeClSampSize)
}
\description{
Calculates proportions of mature fish in successive age classes,
and outputs some variables useful for plotting
}
\examples{
# generate some synthetic maturity data
MinAge = 0
MaxAge = 20
nSexes = 1
nSamples = 300
A50 = 4
A95 = 6
Pmax = 1.0
MaturityParams = c(A50, A95, Pmax)
res=SimulateAgeAtMaturityData(nSamples, nSexes, MinAge, MaxAge, MaturityParams)
ObsAgeCl=res$ObsAgeCl
ObsMatCat=res$ObsMatCat
CalcPropMatureAtAge(MinAge, MaxAge, ObsAgeCl, ObsMatCat)
}
