% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{Calculate_NLL_YearEffectGrowthMod}
\alias{Calculate_NLL_YearEffectGrowthMod}
\title{Calculate NLL associated with fit of year-effects growth model}
\usage{
Calculate_NLL_YearEffectGrowthMod(params)
}
\arguments{
\item{params}{von Bertalanffy growth parameters (some of which are time-varying)}
}
\value{
NLL (and several Report and ADreport variables, associated with RTMB)
}
\description{
This function calculates the negative log-likelihood associated with the fit
of a von Bertalanffy growth model with year-effects parameters (Linf and/or vbK).
The function links to RTMB.
}
\keyword{internal}
