% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{EstPropAtLen_MixtureDistn}
\alias{EstPropAtLen_MixtureDistn}
\title{Return expected proportions at length}
\usage{
EstPropAtLen_MixtureDistn(params)
}
\arguments{
\item{params}{c(Mean1, sd1) normal distn: 1 mode, or
c(Shape1, Rate1) gamma distn: 1 mode, or
c(Mean1, Mean2, sd1, PropZero) normal distn: 2, common sd, or
c(Shape1, Shape2, Rate1, PropZero) gamma distn: 2, common sd, or
c(Mean1, Mean2, sd1, sd2, PropZero) normal distn: 2, separate sds, or
c(Shape1, Shape2, Rate1, Rate2, PropZero) gamma distn: 2 modes, separate sds, or
c(Mean1, Mean2, Mean3, sd1, sd2, sd3, PropZero, PropOne) normal distn: 3 modes, separate sds, or
c(Shape1, Shape2, Shape3, Rate1, Rate2, Rate3, PropZero, PropOne) gamma distn: 3 modes, separate sds}
}
\value{
list object containing estimated proportion at age one, following feasible parameter value
check (EstPropOne), expected frequencies for first, second and third cohorts,
returning NA if not calculated due to specified number of cohorts (Expfreq_mode1, Expfreq_mode2, Expfreq_mode3),
EstPropZero = PropZero, expected overall proportion at length (ExpPropAtLen), expected overall
frequency at length (ExpFreqAtLen), values of penalties for normal disribution models, ensuring
mode 1 is not greater than mode 2, and mode 2 is not greater than mode 3
}
\description{
Return expected proportions at length, for mixture distribution analysis
}
\keyword{internal}
