% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{f2_OffMoveMod}
\alias{f2_OffMoveMod}
\title{Calculate the probability of seeing an observed length (x) within an adult  habitat}
\usage{
f2_OffMoveMod(x, MeanLen, Growthsd, moveL50, moveL95)
}
\arguments{
\item{params}{x, MeanLen, Growthsd, moveL50, moveL95}
}
\value{
returns probability
}
\description{
This function returns a calculated probability of seeing an observed length within an adult habitat, for
use in growth model with offshore movement
}
\keyword{internal}
