% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{f4_OffMoveMod_OffMoveMod}
\alias{f4_OffMoveMod_OffMoveMod}
\title{Calculate the product of x.f(x), for use in calculating the expected length
of x, i.e. the mean, for fish in an adult habitat}
\usage{
f4_OffMoveMod_OffMoveMod(x, MeanLen, Growthsd, moveL50, moveL95)
}
\arguments{
\item{params}{x, MeanLen, Growthsd, moveL50, moveL95}
}
\value{
x.f(x)
}
\description{
This function returns the product of x.f(x), for use in calculating the expected length
of x, i.e. the mean, for fish in a juvenile habitat
}
\keyword{internal}
