% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{FitDivergeGrowthModel}
\alias{FitDivergeGrowthModel}
\title{Fit a von Bertalanffy growth model allowing for divergent growth between sexes or areas}
\usage{
FitDivergeGrowthModel(params, ObsAge, ObsLen, ObsSex)
}
\arguments{
\item{params}{c(log(c(Linf,Linf)),log(c(vbK,vbK)),tzero,log(tdiverge))}

\item{ObsAge}{observed ages}

\item{ObsLen}{observed lengths}

\item{ObsSex}{observed sex}
}
\value{
stored output from internal R nlminb optimisation function (nlmb)
}
\description{
This function fits a von Bertalanffy growth curve, allowing for divergent growth between sexes
or areas after a certain age (tdiverge), to a sample of fish length-at-age data
by minimising the negative log-likelihood associated with the parameters and data, using nlminb
}
\keyword{internal}
