% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{FitLogisticAgeAtMaturityCurve}
\alias{FitLogisticAgeAtMaturityCurve}
\title{Fit a length-based maturity curve to fish maturity-at-length data.}
\usage{
FitLogisticAgeAtMaturityCurve(params, CurveType, nSexes, ObsAgeCl, ObsMatCat)
}
\arguments{
\item{params}{c(A50, A95) ages at 50 and 95 percent maturity}

\item{CurveType}{1 = symmetric logistic, 2 = asymmetric logistic}

\item{nSexes}{number of sexes}

\item{ObsAgeCl}{vector of observed ages}

\item{ObsMatCat}{vector of observed maturity categories (0=immature, 1=mature)}
}
\value{
nlmb (stored output from internal R nlminb optimisation function)
}
\description{
This function fits a logistic maturity curve to a sample of maturity-at-age data
by minimising the negative log-likelihood associated with the parameters and data, using nlminb
}
\keyword{internal}
