% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{FitLogisticLengthAtMaturityCurve}
\alias{FitLogisticLengthAtMaturityCurve}
\title{Fit a length-based maturity curve to fish maturity-at-length data.}
\usage{
FitLogisticLengthAtMaturityCurve(params, CurveType, nSexes, ObsLen, ObsMatCat)
}
\arguments{
\item{params}{c(L50, L95) lengths at 50 and 95 percent maturity}

\item{CurveType}{1 = symmetric logistic, 2 = asymmetric logistic}

\item{nSexes}{number of sexes}

\item{ObsLen}{vector of observed lengths}

\item{ObsMatCat}{vector of observed maturity categories (0=immature, 1=mature)}
}
\value{
nlmb (stored output from internal R nlminb optimisation function)
}
\description{
This function fits a logistic maturity curve to a sample of maturity-at-length data
by minimising the negative log-likelihood associated with the parameters and data, using nlminb
}
\keyword{internal}
