% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{FitMixtureDistnModel}
\alias{FitMixtureDistnModel}
\title{Fit a size mixture distribution model to size frequency data}
\usage{
FitMixtureDistnModel(
  params,
  DistnType,
  Cohorts,
  MinSize,
  MaxSize,
  SizeInt,
  ObsFreq
)
}
\arguments{
\item{params}{c(Mean1, sd1) normal distn: 1 mode, or
c(Shape1, Rate1) gamma distn: 1 mode, or
c(Mean1, Mean2, sd1, PropZero) normal distn: 2, common sd, or
c(Shape1, Shape2, Rate1, PropZero) gamma distn: 2, common sd, or
c(Mean1, Mean2, sd1, sd2, PropZero) normal distn: 2, separate sds, or
c(Shape1, Shape2, Rate1, Rate2, PropZero) gamma distn: 2 modes, separate sds, or
c(Mean1, Mean2, Mean3, sd1, sd2, sd3, PropZero, PropOne) normal distn: 3 modes, separate sds, or
c(Shape1, Shape2, Shape3, Rate1, Rate2, Rate3, PropZero, PropOne) gamma distn: 3 modes, separate sds}

\item{DistnType}{1=normal, 2=gamma}

\item{Cohorts}{1, 2 or 3}

\item{MinSize}{minimum size}

\item{MaxSize}{maximum size}

\item{SizeInt}{size class interval}

\item{ObsFreq}{observed frequency data}
}
\value{
nlmb (stored output from internal R nlminb optimisation function)
}
\description{
This function fits a size mixture distribution model to size frequency data
by minimising the negative log-likelihood associated with the parameters and data, using nlminb.
Function requires size frequency data (stored in memory in R), ObsFreq (Numeric Vector).
}
\keyword{internal}
