% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{FitSchnuteGrowthModel}
\alias{FitSchnuteGrowthModel}
\title{Fit a Schnute growth curve to a sample of fish length-at-age data.}
\usage{
FitSchnuteGrowthModel(
  params,
  nSexes,
  DataType,
  t1,
  t2,
  ObsAge,
  ObsLen,
  ObsMeanLen,
  ObsMeanLense
)
}
\arguments{
\item{params}{c(log(y1),log(y2),a,b) Schnute parameters}

\item{nSexes}{number of sexes}

\item{DataType}{1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis}

\item{t1}{first reference age}

\item{t2}{second reference age}

\item{ObsAge}{observed ages}

\item{ObsLen}{observed lengths}

\item{ObsMeanLen}{mean lengths for specified ages (i.e. as estimated from mixture analysis)}

\item{ObsMeanLense}{se for mean estimated mean lengths for specified ages (i.e. as estimated from mixture analysis)}
}
\value{
nlmb (stored output from internal R nlminb optimisation function)
}
\description{
This function fits a Schnute growth curve to a sample of fish length-at-age data
by minimising the negative log-likelihood associated with the parameters and data, using nlminb.
}
\keyword{internal}
