% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{FitSeasonalGrowthModel}
\alias{FitSeasonalGrowthModel}
\title{Fit a seasonal growth curve to a sample of fish length-at-age data}
\usage{
FitSeasonalGrowthModel(
  params,
  nSexes,
  DataType,
  ObsAge,
  ObsLen,
  ObsMeanLen,
  ObsMeanLense
)
}
\arguments{
\item{params}{Linf, vbK, tzero, tc, C}

\item{nSexes}{number of sexes}

\item{DataType}{1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis}

\item{ObsAge}{observed ages}

\item{ObsLen}{observed lengths}

\item{ObsMeanLen}{mean lengths for specified ages (i.e. as estimated from mixture analysis)}

\item{ObsMeanLense}{se for mean estimated mean lengths for specified ages (i.e. as estimated from mixture analysis)}
}
\value{
nlmb (stored output from internal R nlminb optimisation function)
}
\description{
This function fits a Somers (1988) growth curve to a sample of fish length-at-age data
by minimising the negative log-likelihood associated with the parameters and data, using nlminb
}
\keyword{internal}
