% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{FitvonBertalanffyGrowthModel}
\alias{FitvonBertalanffyGrowthModel}
\title{Fit a von Bertalanffy growth curve to a sample of fish length-at-age data.}
\usage{
FitvonBertalanffyGrowthModel(
  params,
  nSexes,
  DataType,
  ObsAge,
  ObsLen,
  ObsMeanLen,
  ObsMeanLense
)
}
\arguments{
\item{params}{c(log(Linf),log(vbK),tzero) single or combined sex, or c(log(c(Linf,Linf)),log(c(vbK,vbK)),c(tzero,tzero))}

\item{nSexes}{1=single or combined sex, 2=separate sexes}

\item{DataType}{# 1=lengths at age data for individual fish (single sex, or sexes recorded),
2=mean length at age and sd data from mixture analysis, 3=lengths at age data for individual fish
(two sex, sexes not recorded)}

\item{ObsAge}{observed ages}

\item{ObsLen}{observed lengths}

\item{ObsMeanLen}{mean lengths for specified ages (i.e. as estimated from mixture analysis)}

\item{ObsMeanLense}{se for mean estimated mean lengths for specified ages (i.e. as estimated from mixture analysis)}
}
\value{
stored output from internal R nlminb optimisation function (nlmb)
}
\description{
This function fits a von Bertalanffy growth curve to a sample of fish length-at-age data
by minimising the negative log-likelihood associated with the parameters and data, using nlminb
}
\keyword{internal}
