% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{FitvonBertalanffyGrowthModel_prior_param}
\alias{FitvonBertalanffyGrowthModel_prior_param}
\title{Fit a von Bertalanffy growth curve to a sample of fish length-at-age data, with prior imposed for tzero.}
\usage{
FitvonBertalanffyGrowthModel_prior_param(
  params,
  nSexes,
  DataType,
  ObsAge,
  ObsLen,
  tzero_mean,
  tzero_sd,
  Linf_mean,
  Linf_sd
)
}
\arguments{
\item{params}{c(log(Linf),log(vbK),tzero) single or combined sex, or c(log(c(Linf,Linf)),log(c(vbK,vbK)),c(tzero,tzero))}

\item{nSexes}{1=single or combined sex, 2=separate sexes}

\item{DataType}{# 1=lengths at age data for individual fish (single sex, or sexes recorded),
2=mean length at age and sd data from mixture analysis, 3=lengths at age data for individual fish
(two sex, sexes not recorded)}

\item{ObsAge}{observed ages}

\item{ObsLen}{observed lengths}

\item{tzero_mean}{mean value for tzero (user specified)}

\item{tzero_sd}{standard deviation for tzero (user specified)}
}
\value{
stored output from internal R nlminb optimisation function (nlmb)
}
\description{
This function fits a von Bertalanffy growth curve to a sample of fish length-at-age data
by minimising the negative log-likelihood associated with the parameters and data, using nlminb,
including a normal penalty prior for tzero.
}
\keyword{internal}
