% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{GetConfidenceLimitsForMaturityCurve}
\alias{GetConfidenceLimitsForMaturityCurve}
\title{Get confidence limits for the fitted maturity curve}
\usage{
GetConfidenceLimitsForMaturityCurve(
  params,
  vcov.params,
  CurveType,
  nSexes,
  LogisticModType,
  plotlengthrange,
  plotages
)
}
\arguments{
\item{params}{c(L50,L95) or c(A50, A95) length or age-based logistic maturity parameters}

\item{vcov.params}{estimated variance-covariance matrix for maturity parameters}

\item{CurveType}{1=symmetric logistic, 2=asymmetric logistic}

\item{nSexes}{number of sexes}

\item{LogisticModType}{1=length-based, 2=age-based}

\item{plotlengthrange}{lower and upper length class range for plotting lengths}

\item{plotages}{ages for plotting if age based (set to NA is length-based)}
}
\value{
scatter plot of length-at-age data
}
\description{
Function uses parametric resample to produce 95 percent confidence limits for a fitted maturity curve
}
