% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{GetConfidenceLimitsForMixtureDistnCurve}
\alias{GetConfidenceLimitsForMixtureDistnCurve}
\title{Get expected size frequency from size mixture distribution model}
\usage{
GetConfidenceLimitsForMixtureDistnCurve(params, vcov.params, DistnType)
}
\arguments{
\item{params}{c(Mean1, sd1) normal distn: 1 mode, or
c(Shape1, Rate1) gamma distn: 1 mode, or
c(Mean1, Mean2, sd1, PropZero) normal distn: 2, common sd, or
c(Shape1, Shape2, Rate1, PropZero) gamma distn: 2, common sd, or
c(Mean1, Mean2, sd1, sd2, PropZero) normal distn: 2, separate sds, or
c(Shape1, Shape2, Rate1, Rate2, PropZero) gamma distn: 2 modes, separate sds, or
c(Mean1, Mean2, Mean3, sd1, sd2, sd3, PropZero, PropOne) normal distn: 3 modes, separate sds, or
c(Shape1, Shape2, Shape3, Rate1, Rate2, Rate3, PropZero, PropOne) gamma distn: 3 modes, separate sds}

\item{vcov.params}{estunated variance-covariance matrix for parameters of mixture model}

\item{DistnType}{1=normal, 2=gamma}
}
\value{
median values and lower and upper 95 percent confidence limits for expected size frequencies
at length calculated from parametric resampling procedure (sim.size.est, sim.size.low, sim.size.up),
median values and lower and upper 95 percent confidence limits for gamma mean and modal values,
from parametric resampling procedure, giving NAs if normal distribution specified, and if a parameter
is not required for specified model with only 1 or 2 modes (gamma_mean1, gamma_mean2,
gamma_mean3, gamma_mode1, gamma_mode2, gamma_mode3, gamma_mean1_lw, gamma_mean2_lw, gamma_mean3_lw,
gamma_mode1_lw, gamma_mode2_lw, gamma_mode2_lw, gamma_mean1_up, gamma_mean2_up, gamma_mean3_up,
gamma_mode1_up, gamma_mode2_up, gamma_mode3_up)
}
\description{
Gets expected size frequency from size mixture distribution model, used in calculating
confidence limits for mixture distribution curves applying a parametric resampling procedure
}
\keyword{internal}
