% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{GetDivergeGrowthModelResults}
\alias{GetDivergeGrowthModelResults}
\title{Get statistical outputs from a fitted von Bertalanffy growth curve.}
\usage{
GetDivergeGrowthModelResults(params, ObsAge, ObsLen, ObsSex)
}
\arguments{
\item{params}{c(log(c(Linf,Linf)),log(c(vbK,vbK)),tzero,log(tdiverge))}

\item{ObsAge}{observed ages}

\item{ObsLen}{observed lengths}

\item{ObsSex}{observed sex (1=female, 2=male)}
}
\value{
nll, convergence, SampleSize ParamEst, PtEst_LdivergePtEst, PtEst_Maltzero, params, vcov.params, cor.params
}
\description{
This function fits a von Bertalanffy growth curve to a sample of fish length-at-age data
by minimising the negative log-likelihood associated with the parameters and data,
using nlminb. It provides various statistical outputs in include convergence statistics,
parameter estimated and associated 95% confidence limits and associated variance-covariance matrix,
calculated using the MASS package
}
\examples{
set.seed(123)
GrowthEqn = 4 # von Bertalanffy growth equation with divergent growth
nSexes = 2 # single or combined sex
nSamples = 500
MinAge = 0
MaxAge = 20
AgeStep = 1
Linf = c(1000,500)
vbK = c(0.15,0.15)
tzero1 = 0
tdiverge = 1
Growth_params = c(Linf, vbK, tzero1, tdiverge)
Ref_ages = NA
Growth_cv = 0.1
Res = SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
x=which(Res$ObsDat$ObsSex==1)
plot(Res$ObsAge[x], Res$ObsLen[x])
y=which(Res$ObsDat$ObsSex==2)
points(Res$ObsAge[y], Res$ObsLen[y],col="blue")
# fit growth model
InitLinf = c(1050,500)
# InitvbK = c(0.13,0.13) # separate vbK option
InitvbK = 0.13 # single vbK option
Inittzero1 = 0
Inittdiverge = 4
params = c(log(InitLinf), log(InitvbK), Inittzero1, log(Inittdiverge))
ObsAge = Res$ObsDat$ObsAge
ObsLen = Res$ObsDat$ObsLen
ObsSex = Res$ObsDat$ObsSex
FittedRes=GetDivergeGrowthModelResults(params, ObsAge, ObsLen, ObsSex)
FittedRes$Boot_ParamEst
}
