% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{GetLogisticMaturityCurveResultsErrOpt1}
\alias{GetLogisticMaturityCurveResultsErrOpt1}
\title{Fit logistic maturity curve and get parameter estimates and uncertainty}
\usage{
GetLogisticMaturityCurveResultsErrOpt1(
  params,
  nSexes,
  LogisticModType,
  CurveType,
  ObsLen,
  ObsAgeCl,
  ObsMatCat
)
}
\arguments{
\item{params}{c(L50,L95) length at 50 and 95 percent maturity (L50)}

\item{nSexes}{c(L50,L95) number of sexes}

\item{LogisticModType}{1=length-based, 2=age-based}

\item{CurveType}{1=symmetric logistic, 2=asymmetric logistic}

\item{ObsLen}{vector of observed lengths (set to NA for age model)}

\item{ObsAgeCl}{vector of observed age classes (set to NA for length model)}

\item{ObsMatCat}{vector of observed maturity categories (0=immature, 1=mature)}
}
\value{
negative log-likelihood (nll), nlminb convergence diagnostic (convergence), sample size (SampleSize)
maturity parameter estimates with lower and upper 95 percent confidence limits (ParamEst), point estimates
for growth parameters (params), variance-covariance matrix (vcov.params)
}
\description{
This function fits a logistic maturity curve to a sample of maturity-at-length data
by minimising the negative log-likelihood associated with the parameters and data, using nlminb.
Estimates of uncertainty are based on asymtotic error approximations from estimated variance
covariance matrix
}
