% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{GetMixtureModelResults}
\alias{GetMixtureModelResults}
\title{Get statistical outputs from a fitted size mixture distribution model}
\usage{
GetMixtureModelResults(
  params,
  DistnType,
  Cohorts,
  MinSize,
  MaxSize,
  SizeInt,
  ObsFreq
)
}
\arguments{
\item{params}{c(Mean1, sd1) normal distn: 1 mode, or
c(Shape1, Rate1) gamma distn: 1 mode, or
c(Mean1, Mean2, sd1, PropZero) normal distn: 2, common sd, or
c(Shape1, Shape2, Rate1, PropZero) gamma distn: 2, common sd, or
c(Mean1, Mean2, sd1, sd2, PropZero) normal distn: 2, separate sds, or
c(Shape1, Shape2, Rate1, Rate2, PropZero) gamma distn: 2 modes, separate sds, or
c(Mean1, Mean2, Mean3, sd1, sd2, sd3, PropZero, PropOne) normal distn: 3 modes, separate sds, or
c(Shape1, Shape2, Shape3, Rate1, Rate2, Rate3, PropZero, PropOne) gamma distn: 3 modes, separate sds}

\item{DistnType}{1=normal, 2=gamma}

\item{Cohorts}{1, 2 or 3}

\item{MinSize}{minimum size}

\item{MaxSize}{maximum size}

\item{SizeInt}{size class interval}

\item{ObsFreq}{observed frequency data}
}
\value{
returns a list object containing negative log-likelihood (nll), nlminb convergence diagnostic
(convergence), sample size (SampleSize), estimates of model parameters with lower and upper 95
confidence limits (ParamEst), Estimates of standard deviations for estimated mean
(normal distribution assumption) or shape parameters (gamma distribution assumption) for
each fitted size mode (EstMean1_sd, EstMean2_sd, EstMean3_sd), point estimates
for model parameters (params) and variance-covariance matrix (vcov.params),
expected frequency for cohorts 1, 2 and 3 (expfreq1, expfreq2, expfreq3),
overall expected proportions at length across all cohorts (ExpPropAtLen), and
overall expected frequencies at length,across all cohorts (ExpFreqAtLen).
}
\description{
This function fits a size mixture distribution model to size frequency data
by minimising the negative log-likelihood associated with the parameters and data, using nlminb.
It provides various outputs in include convergence statistics, parameter estimated
and associated 95 calculated using the MASS package.
}
\examples{
Simulate data with 1 mode, specifying normal distribution
set.seed(123)
MinSize = 0
MaxSize = 60
SizeInt = 1
SampSize = 1000
Cohorts=1
Mean1 = 20
sd1 = 5
ObsSize = round(rnorm(SampSize, Mean1, sd1),0)
HistData=hist(ObsSize, breaks=seq(0,MaxSize,SizeInt), right=FALSE, plot=FALSE)
ObsFreq = HistData$counts
# Specify starting parameter values for Mean1 and sd1, and return associated negative log-likelihood
DistnType = 1 # 1=normal, 2=gamma
params = log(c(30, 5)) # normal
GetMixtureModelResults(params, DistnType, Cohorts, MinSize, MaxSize, SizeInt, ObsFreq)
# Simulate data with 1 mode, specifying gamma distribution
set.seed(123)
MinSize = 0
MaxSize = 60
SizeInt = 1
SampSize = 1000
Cohorts=1
Shape1 = 20
Rate1 = 1
ObsSize = round(rgamma(SampSize, Shape1, Rate1),0)
HistData=hist(ObsSize, breaks=seq(0,MaxSize,SizeInt), right=FALSE, plot=FALSE)
ObsFreq = HistData$counts
DistnType = 2 # 1=normal, 2=gamma
params = log(c(20, 1)) # gamma
res=GetMixtureModelResults(params, DistnType, Cohorts, MinSize, MaxSize, SizeInt, ObsFreq)
# Get mean and mode of fitted gamma distribution
alpha = res$ParamEst[1,1]
beta = 1 / res$ParamEst[2,1]
# Simulate data with 2 modes, specifying normal distribution
set.seed(123)
MinSize = 0
MaxSize = 60
SizeInt = 1
SampSize = 1000
# normal
Mean1 = 20
Mean2 = 35
sd1 = 5
sd2 = 5
PropZero = 0.7
ObsSize1 = round(rnorm(PropZero*SampSize, Mean1, sd1),0)
ObsSize2 = round(rnorm((1-PropZero)*SampSize, Mean2, sd2),0)
ObsSize = c(ObsSize1,ObsSize2)
HistData=hist(ObsSize, breaks=seq(0,MaxSize,SizeInt), right=FALSE, plot=FALSE)
ObsFreq = HistData$counts
params = c(log(c(15, 40, 5, 5)), 0.5) # separate sd's for the 2 cohorts
Cohorts=2
DistnType = 1 # 1=normal, 2=gamma
GetMixtureModelResults(params, DistnType, Cohorts, MinSize, MaxSize, SizeInt, ObsFreq)
# Simulate data with 2 modes, specifying gamma distribution
set.seed(123)
MinSize = 0
MaxSize = 60
SizeInt = 1
SampSize = 1000
Shape1 = 20
Shape2 = 35
Rate1 = 1.5
Rate2 = 1
PropZero = 0.7
ObsSize1 = round(rgamma(PropZero*SampSize, Shape1, Rate1),0)
ObsSize2 = round(rgamma((1-PropZero)*SampSize, Shape2, Rate2),0)
ObsSize = c(ObsSize1,ObsSize2)
HistData=hist(ObsSize, breaks=seq(0,MaxSize,SizeInt), right=FALSE, plot=TRUE)
ObsFreq = HistData$counts
params = c(log(c(15, 40, 1, 1)), 0.5) # separate sd's for the 2 cohorts
Cohorts=2
DistnType = 2 # 1=normal, 2=gamma
res=GetMixtureModelResults(params, DistnType, Cohorts, MinSize, MaxSize, SizeInt, ObsFreq)
# Get means and modes of fitted gamma distributions
alpha = res$ParamEst[1,1]
beta = 1 / res$ParamEst[3,1]
Dist_mean1 = alpha * beta
Dist_mode1 = (alpha - 1) * beta
alpha = res$ParamEst[2,1]
beta = 1 / res$ParamEst[4,1]
Dist_mean2 = alpha * beta
Dist_mode2 = (alpha - 1) * beta
}
