% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{GetSpawningDurationModelResults}
\alias{GetSpawningDurationModelResults}
\title{Get results for fitted spawning duration model}
\usage{
GetSpawningDurationModelResults(
  params,
  FishLen,
  DecDay,
  ObsMatStatus,
  MinLen,
  MaxLen,
  LenInc,
  nsims
)
}
\arguments{
\item{params}{specified (or estimated), transformed parameter values, for the spawning duration model (including
lnL50, lnslope, logitPkSpawn, lnkappa, lnkappa2, lnslope1, lnslope2)}

\item{FishLen}{Fish lengths}

\item{DecDay}{Days of year of fish capture (in decimal years)}

\item{ObsMatStatus}{maturity status, 0=immature, 1=mature}

\item{MinLen}{Minimum fish length (for determining lengths to output expected model curves)}

\item{MaxLen}{Maximum fish length for analysis}

\item{LenInc}{Length increment (for determining lengths to output expected model curves)}

\item{nsims}{number of samples to generate using parametric resampling. Set to zero to turn off resampling.}
}
\value{
estimated parameters (params), negative log-likelihood of fitted model (NLL),
convergence statistic (0=converged), estimated variance-covariance matrix for estimated parameters,
estimated parameters with 95 percent asymptotic confidence limits (including back-transformed parameters),
estimated parameters with 95 percent confidence limits dervied from parametric resampling of estimated
parameters and associated variance-covariance matrix (including back-transformed parameters),
}
\description{
This function fits a model for estimating spawning season timing and duration, allowing for
fish length effects. The curves associated with the fitted model, for fish of a given length,
have the appearance are approximately bell-shaped, and can vary in heights among fish of different
lengths. The model may also described as an asymmetric double logistic model with variable height.
}
\examples{
# Simulate spawning proportion data
# First, simulate length and age data
set.seed(123)
GrowthEqn=1 # von Bertalanffy
nSamples = 5000
nSexes = 1
MinAge = 1
MaxAge = 20
AgeStep = 1
Ref_ages = NA
Linf = 1000
vbK = 0.1
tzero = 0
Growth_params = c(Linf,vbK,tzero)
Growth_cv = 0.08
# Spawning duration model parameters
lnL50 = log(300)
lnslope= log(0.015)
logitPkSpawn = log(0.6/(1-0.6))
lnkappa = log(0.0002)
lnkappa2 = log(0.0002)
lnslope1 = log(30)
lnslope2 = log(30)
params = c(lnL50,lnslope,logitPkSpawn,lnkappa,lnkappa2,lnslope1,lnslope2)
Res=SimulateSpawningDurationData(params, GrowthEqn, nSamples, nSexes, MinAge, MaxAge,
                                 AgeStep, Ref_ages, Growth_params, Growth_cv)
ObsMatStatus=Res$ObsMatStatus
FishLen=Res$FishLen
DecDay=Res$DecDay
# Spawning duration model parameters - initial values
lnL50 = log(350)
lnslope= log(0.02)
logitPkSpawn = log(0.7/(1-0.7))
lnkappa = log(0.00015)
lnkappa2 = log(0.00035)
lnslope1 = log(25)
lnslope2 = log(35)
params = c(lnL50,lnslope,logitPkSpawn,lnkappa,lnkappa2,lnslope1,lnslope2)
nsims=20
MinLen=0
MaxLen=1100
LenInc=100
res=GetSpawningDurationModelResults(params, FishLen, DecDay, ObsMatStatus, MinLen, MaxLen, LenInc, nsims)
}
