% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{GetvonBertalanffyGrowthResults_prior_param}
\alias{GetvonBertalanffyGrowthResults_prior_param}
\title{Get statistical outputs from a fitted von Bertalanffy growth curve, with imposed normal prior for tzero.}
\usage{
GetvonBertalanffyGrowthResults_prior_param(
  params,
  nSexes,
  DataType,
  ObsAge,
  ObsLen,
  tzero_mean,
  tzero_sd
)
}
\arguments{
\item{params}{c(log(Linf),log(vbK),tzero) single or combined sex, or c(log(c(Linf,Linf)),log(c(vbK,vbK)),c(tzero,tzero))}

\item{nSexes}{1=single or combined sexes, 2=separate sexes}

\item{DataType}{# 1=lengths at age data for individual fish (single sex, or sexes recorded),
2=mean length at age and sd data from mixture analysis, 3=lengths at age data for individual fish
(two sex, sexes not recorded)}

\item{ObsAge}{observed ages}

\item{ObsLen}{observed lengths}

\item{ObsMeanLen}{mean lengths for specified ages (i.e. as estimated from mixture analysis)}

\item{ObsMeanLense}{se for mean estimated mean lengths for specified ages (i.e. as estimated from mixture analysis)}
}
\value{
negative log-likelihood (nll), nlminb convergence diagnostic (convergence)
sample size (SampleSize), growth parameter estimates with lower and upper 95%
confidence limits (ParamEst), point estimates for growth parameters (params)
and variance-covariance matrix (vcov.params)
}
\description{
This function fits a von Bertalanffy growth curve to a sample of fish length-at-age data
by minimising the negative log-likelihood associated with the parameters and data,
using nlminb. A normal prior is imposed for estimating tzero.
It provides various statistical outputs in include convergence statistics,
parameter estimated and associated 95% confidence limits and associated variance-covariance matrix,
calculated using the MASS package
}
\examples{
# ***************************
# simulate length-at-age data
# ***************************
library(WAFishBiology)
library(L3Assess)
set.seed(123)
SampleSize=500 # sample size for retained catches (and same number for released fish, if an MLL is specified)
MaxAge = 80
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
NatMort = 0.05
FishMort = 1.5*NatMort
MaxLen = 1000
LenInc = 20
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified
SelectivityType=2 # 1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve
SelectivityAtLen = NA # selectivity vector
SelParams = c(450, 10) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0 # proportion of fish that die due to natural mortality
# single sex, von Bertalanffy
GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
Linf = 600
vbK = 0.2
CVSizeAtAge = 0.06
GrowthParams = c(Linf, vbK)
RefnceAges = NA
SimRes=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                                  SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
# plot data
PlotOpt=1 # 0=all plots, 1=retained lengths at age and growth curves, 2=retained plus discard lengths at age
# and growth curves, 3=retained plus discard length frequency, 4=retained plus discard age frequency,
# 5=sex specific retained lengths at age and growth curves, 6=retained plus discard lengths at age and growth curves,
# 7=female length frequency, 8=male length frequency, 9=female age frequency, 10=male age frequency,
# 11=selectivity/retention, 12=F-at-age reten + disc, 13=F-at-age reten, 14=F-at-age disc
PlotSimLenAndAgeFreqData_EqMod(MaxAge, MaxLen, SimRes, PlotOpt)
# ************************
# fit traditional vb model
# ************************
ObsAge = as.vector(SimRes$ObsDecAgeRetCatch)
ObsLen = as.vector(SimRes$ObsRandLenRetCatch)
plot(ObsAge,ObsLen, xlim=c(0,80), ylim=c(0,1000))
DataType=1  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
nSexes=1
params = c(log(400),log(0.3),0) # log(Linf), log(k), tzero
FittedRes=GetvonBertalanffyGrowthResults(params, nSexes, DataType, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLense=NA)
FittedRes$ParamEst
plotages=seq(0, MaxAge,0.1)
GrowthEqn = 1
par(mfrow=c(2,2))
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen,
                      ObsMeanLense, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
# **********************************************
# fit traditional vb model with tzero constraint
# **********************************************
DataType=1  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
nSexes=1
params = c(log(400),log(0.3),0) # log(Linf), log(k), tzero
tzero_mean=-1.5 # specify prior mean
tzero_sd=1.0 # specify prior sd
FittedRes=GetvonBertalanffyGrowthResults_prior_param(params, nSexes, DataType, ObsAge, ObsLen, tzero_mean, tzero_sd)
FittedRes$ParamEst
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen,
                      ObsMeanLense, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
# posterior
Esttzero_mean = unlist(FittedRes$par[3])
vcov.params = FittedRes$vcov.params
ses = sqrt(diag(vcov.params))
Esttzero_sd = unlist(ses[3])
# plot to compare prior vs posterior - scaled density plot
prior_data <- rnorm(1000,tzero_mean,tzero_sd)
posterior_data <- rnorm(1000,Esttzero_mean,Esttzero_sd)
hist(prior_data, breaks = seq(-10,3,0.2), freq = FALSE,
     col = NA, border = NA, xlim = c(-5,1), ylim=c(0,1),
     xlab = "Value", ylab = "Density", main="tzero", cex.main=0.8)
lines(density(prior_data),     col = "blue", lwd = 2)
lines(density(posterior_data), col = "red",  lwd = 2)
legend("topright", legend = c("Prior", "Posterior"),
       lwd=2, col=c("blue","red"),
       border = NA, bty='n')
}
