% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{GetWeightLengthRegressionResults}
\alias{GetWeightLengthRegressionResults}
\title{Fit weight-length relationship and get parameter estimates}
\usage{
GetWeightLengthRegressionResults(FishLen, FishWt)
}
\arguments{
\item{FishLen}{vector of observed fish lengths}

\item{FishWt}{vector of observed fish weights}
}
\value{
outputs of linear model used to fit weight-length relationship in log space (mod1), weight-length parameters
with associated 95 percent confidence limits (ParamEst), adjusted coefficient of determination for fitted model, R2
(adj.r.squared), lengths for plotting line in log space (plot_lnlen), weights for plotting line in log space (plot_lnwt),
95 percent confidence limits for fitted line (conf_int), 95 percent prediction intervals (pred_int), residual variance (Resid_var)
}
\description{
Function fits a weight-length relationship to data in log space using the lm() functin in R
}
\examples{
# generate synthetic data
nsamples = 100
minlen = 20
maxlen = 500
lenwt_a = 0.00002
lenwt_b = 3
lenwt_sd = 0.1
Res = SimulateWeightLengthData(nsamples, minlen, maxlen,
                               lenwt_a, lenwt_b, lenwt_sd)
FishLen=Res$FishLen
FishWt=Res$FishWt
# Fit weight-length relationship and get results
GetWeightLengthRegressionResults(FishLen, FishWt)
}
