% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{GetYearEffectGrowthModResults}
\alias{GetYearEffectGrowthModResults}
\title{Fit von Betalanffy growth model with year-effects parameters}
\usage{
GetYearEffectGrowthModResults(dat, params)
}
\arguments{
\item{params}{von Bertalanffy growth parameters (some of which are time-varying)}
}
\value{
SummaryRes (par, convergence, NLL, sdr, params_pt_est, params_sd=params_sd, ParamEst=ParamEst),
RTMBReport (EstLenAtIntAge, ObsLenAtAge)
}
\description{
This function fits a von Bertalanffy growth model with year-effects parameters (Linf and/or vbK).
The function uses RTMB for optimisation.
}
\examples{
library(L3Assess)
library(RTMB)
set.seed(123)
MinAge = 1
MaxAge = 40
Ages = MinAge:MaxAge
NatMort = 0.11
FMort = 0.11
ZMort = FMort + NatMort
SelA50 = c(7,6)
SelA95 = c(8,7)
SampleSize = 150 # required sample size. For 2 sex model, same sample size generated for each sex.
Res=SimAgeFreqData_EqMod(SampleSize, MinAge, MaxAge, SelA50, SelA95, NatMort, FMort)
ObsAgeFreq = Res$CatchSample
par(mfrow=c(1,1),mar=c(4,3,2,2))
plot(MinAge:MaxAge, ObsAgeFreq[1,],col="red",'o')
lines(MinAge:MaxAge, ObsAgeFreq[2,],col="blue")
# Specify information required to generate length-at-age data with
# time-varying growth
StartYear = 2000
EndYear = 2010
nYrs = EndYear - StartYear + 1
minIntAge = 0
MaxIntAge = 40
nSexes = 2
# Specify true growth parameter values. Linf and/or vbK can be year-specific.
nLinfVals = nYrs
nvbKVals = nYrs
Linf_Fem = seq(1000,900,-10)
Linf_Mal = seq(1100,1000,-10)
Linf = data.frame(data.frame(Linf_Fem=Linf_Fem,Linf_Mal=Linf_Mal))
vbK_Fem = seq(0.2,0.1,-0.01)
vbK_Mal = seq(0.2,0.1,-0.01)
vbK = data.frame(data.frame(vbK_Fem=vbK_Fem,vbK_Mal=vbK_Mal))
tzero = 0 # Specify common tzero for both sexes to simplify
CV = 0.1
Res = SimulateTimeVaryingLengthAtAgeData(Linf, vbK, tzero, CV, nLinfVals, nvbKVals, StartYear, EndYear, minIntAge, MaxIntAge, nSexes)
ObsLenAtAgeDat = Res$ObsLenAtAgeDat
# Fit year-effects growth model in RTMB
# Specify starting values
InitLinf_Fem = rep(1000,nLinfVals)
InitLinf_Mal = rep(1100,nLinfVals)
InitvbK_Fem = rep(0.12,nvbKVals)
InitvbK_Mal = rep(0.13,nvbKVals)
Inittzero = 0
# RTMB data input list
dat <- list(nLinfVals=nLinfVals,
            nvbKVals=nvbKVals,
            ObsLenAtAgeDat=ObsLenAtAgeDat)
# RTMB param input list
params <- list(lnLinf_Fem=log(InitLinf_Fem),
               lnLinf_Mal=log(InitLinf_Mal),
               lnvbK_Fem=log(InitvbK_Fem),
               lnvbK_Mal=log(InitvbK_Mal),
               tzero=Inittzero)
FittedRes = GetYearEffectGrowthModResults(dat, params)
Age = ObsLenAtAgeDat$Age
RandLenAtAge = ObsLenAtAgeDat$RandLenAtAge
YrIndex = ObsLenAtAgeDat$YrIndex
Sex = ObsLenAtAgeDat$Sex
IntAge = floor(ObsLenAtAgeDat$Age)
# plot simulated length at age data
for (y in 1:nYrs) {
  for (s in 1:nSexes) {
    x=which(YrIndex==y & Sex==s)
    if (y==1 & s==1) {
      plot(Age[x], RandLenAtAge[x], xlim=c(0,MaxAge), ylim=c(0,1400), col="red")
    }
    if (y>1 & s==1) {
      points(Age[x], RandLenAtAge[x], col="red")
    }
    if (s==2) {
      points(Age[x], RandLenAtAge[x], col="blue")
    }
  }
}
# plot female data for some specific ages
par(mfcol=c(4,2), mar=c(4,3,2,2))
for (s in 1:2) {
  for (a in 7:10) {
    x=which(Sex==s & IntAge==a)
    colour = c("red","blue")
    plot(YrIndex[x], RandLenAtAge[x], xlim=c(1,nYrs), ylim=c(300,1200), col=colour[s],main=paste("Age =",a))
    lines(1:nYrs,  FittedRes$RTMBReport$EstLenAtIntAge[,s,a])
  }
}
}
