% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{ilogit}
\alias{ilogit}
\title{Inverse logit transformation}
\usage{
ilogit(x)
}
\arguments{
\item{x}{logit-transformed value}
}
\value{
inverse of logit-transformed value
}
\description{
Inverse logit transformation (back-transforms logit-transformed values, used
to keep parameters between 0 and 1 when fitting a model)
}
\keyword{internal}
