% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{LogisticEqnAgeAtMaturity}
\alias{LogisticEqnAgeAtMaturity}
\title{Logistic curve for age at maturity}
\usage{
LogisticEqnAgeAtMaturity(params)
}
\arguments{
\item{params}{c(A50, A95) or c(Pmax, A50, A95) for CurveType=1, or c(Q, B, V) or c(Pmax, Q, B, V) for CurveType=2}
}
\value{
probability of maturity for observed ages. Used in calculation of objective function
for length at maturity model
}
\description{
This function applies a logistic curve (symmetric or asymmetric) for describing probability of maturity at age, used
in calculation of objective function for maturity model
}
\keyword{internal}
