% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{LogisticEqnAgeAtMaturity2}
\alias{LogisticEqnAgeAtMaturity2}
\title{Logistic curve for age at maturity (for plotting)}
\usage{
LogisticEqnAgeAtMaturity2(params, CurveType, nSexes, plotages)
}
\arguments{
\item{params}{c(A50, A95) or c(Pmax, A50, A95) for CurveType=1, or c(Q, B, V) or c(Pmax, Q, B, V) for CurveType=2}

\item{CurveType}{number of sexes}

\item{nSexes}{number of sexes}

\item{plotages}{age classes for plotting}
}
\value{
probability of maturity for observed ages. Used in calculation of objective function
for length at maturity model
}
\description{
This function applies a logistic curve for describing probability of maturity at age, used for plotting
maturity results
}
\keyword{internal}
