% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{logit}
\alias{logit}
\title{Logit transformation}
\usage{
logit(x)
}
\arguments{
\item{x}{value to be transformed}
}
\value{
logit-transformed value
}
\description{
Logit-transformation, used to keep parameters between 0 and 1 when fitting a model
}
\keyword{internal}
