% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{penfun}
\alias{penfun}
\title{Smooth penalty function penalty value calculation}
\usage{
penfun(x, eps, pen)
}
\arguments{
\item{x}{value to be check for penalty constraint}
}
\value{
value of x after checking for constraint
}
\description{
Smooth penalty function, equivalent functionality to ADMB posfun (excluding use of if statements)
}
\keyword{internal}
