% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{PlotAgeAtMaturityData}
\alias{PlotAgeAtMaturityData}
\title{Plot observed fish maturity-at-age data}
\usage{
PlotAgeAtMaturityData(
  nSexes,
  AgeCl,
  PropMat,
  AgeClSampSize,
  xmax,
  xint,
  GraphTitle,
  SampSizelab_cex,
  xaxis_lab,
  yaxis_lab
)
}
\arguments{
\item{nSexes}{number of sexes}

\item{AgeCl}{vector of observed age classes}

\item{PropMat}{vector of calculated proportion mature by age class}

\item{AgeClSampSize}{vector of age class sample sizes}

\item{xmax}{x axis maximum}

\item{xint}{x axis interval}

\item{GraphTitle}{graph title}

\item{SampSizelab_cex}{sample size graph label size}

\item{xaxis_lab}{x axis label}

\item{yaxis_lab}{y axis label}
}
\value{
scatter plot of maturity-at-age data
}
\description{
This function produces a scatter plot plot of fish age-at-maturity data
}
\examples{
# generate some synthetic age-at-maturity data and plot
# age-at-maturity data
MinAge = 0
MaxAge = 20
# nSexes = 1
# nSamples = 300
# A50 = 4
# A95 = 6
# Pmax = 1.0
# MaturityParams = c(A50, A95, Pmax)
nSexes = 2
nSamples = c(300,300)
A50 = c(4,4.5)
A95 = c(6,6)
Pmax = c(1.0,1.0)
MaturityParams = c(A50, A95, Pmax)
res=SimulateAgeAtMaturityData(nSamples, nSexes, MinAge, MaxAge, MaturityParams)
AgeCl = res$AgeClasses
# Single sex
# PropMat = res$PropMat
# AgeClSampSize=res$AgeClSampSize
# separate sexes
PropMat = t(data.frame(FemPropMat=res$FemPropMat,MalPropMat=res$MalPropMat))
AgeClSampSize = t(data.frame(FemAgeClSampSize=res$FemAgeClSampSize,MalAgeClSampSize=res$MalAgeClSampSize))
xmax=MaxAge
xint=2
GraphTitle=NA
xaxis_lab=NA
yaxis_lab=NA
SampSizelab_cex=NA
PlotAgeAtMaturityData(nSexes, AgeCl, PropMat, AgeClSampSize, xmax, xint,
                      GraphTitle, SampSizelab_cex, xaxis_lab, yaxis_lab)
}
