% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{PlotFittedAgeAtMaturityCurve}
\alias{PlotFittedAgeAtMaturityCurve}
\title{Plot fitted logistic curve to fish maturity-at-age data.}
\usage{
PlotFittedAgeAtMaturityCurve(
  params,
  CurveType,
  nSexes,
  ObsAgeCl,
  ObsMatCat,
  AgeCl,
  PropMat,
  AgeClSampSize,
  plotages,
  xmax,
  xint,
  GraphTitle,
  xaxis_lab,
  yaxis_lab,
  SampSizelab_cex,
  ShowSampSizes,
  ShowLegend,
  PlotCLs
)
}
\arguments{
\item{params}{age-based logistic maturity curve parameters}

\item{CurveType}{1=symmetric, 2=asymmetric}

\item{nSexes}{number of sexes}

\item{ObsAgeCl}{vector of observed age classes}

\item{ObsMatCat}{vector of observed maturity categories (0=immature, 1=mature)}

\item{AgeCl}{vector of observed age classes}

\item{PropMat}{proportion mature in age classes}

\item{AgeClSampSize}{vector of observed sample sizes for observed age classes}

\item{plotages}{ages for plotting}

\item{xmax}{x axis maximum}

\item{xint}{x axis interval}

\item{GraphTitle}{graph title}

\item{xaxis_lab}{x axis label}

\item{yaxis_lab}{y axis label}

\item{SampSizelab_cex}{font size for sample size labels}

\item{ShowSampSizes}{logical (set to TRUE to show sample size labels)}

\item{PlotCLs}{logical (set to TRUE to show 95 percent confidence limits for curve)}
}
\description{
This function fits a logistic maturity curve to maturity-at-age data. Various plotting
options provided.
}
\examples{
# plot model fitted to single sex/combined sexes
# generate synthetic data
MinAge = 0
MaxAge = 20
nSexes = 1
nSamples = 500
CurveType = 1
A50 = 4
A95 = 6
Pmax = 1.0
MaturityParams = c(A50, A95, Pmax)
res=SimulateAgeAtMaturityData(nSamples, CurveType, nSexes, MinAge, MaxAge, MaturityParams)
AgeCl = res$AgeClasses
PropMat = res$PropMat
AgeClSampSize=res$AgeClSampSize
xmax=MaxAge
xint=2
GraphTitle=NA
xaxis_lab=NA
yaxis_lab=NA
ObsLen=NA
ObsAgeCl=res$ObsAgeCl
ObsMatCat=res$ObsMatCat
LogisticModType = 2 # 1=length, 2=age
res$ParamEst
plotages=res$plotages
# 2 parameter model
InitA50 = 3
InitA95 = 5
params = c(InitA50, InitA95) # without Pmax parameter
# # 3 parameter model
# InitA50 = 4.5
# InitA95 = 7
# InitPmax = 0.9
# InitPmax_logit = log(InitPmax/(1-InitPmax)) # logit transform
# params = c(InitPmax_logit, InitA50, InitA95) # with Pmax parameter
PlotFittedAgeAtMaturityCurve(params, CurveType, nSexes, ObsAgeCl, ObsMatCat, AgeCl, PropMat, AgeClSampSize, plotages,
                             xmax=20, xint=2, GraphTitle=NA, xaxis_lab=NA,
                             yaxis_lab=NA, SampSizelab_cex = NA, ShowSampSizes=FALSE, ShowLegend=FALSE, PlotCLs=FALSE)
# plot model fitted to separate sexes (simultaneously)
# generate synthetic data
# MinAge = 0
# MaxAge = 20
# nSexes = 2
# nSamples = c(500,500)
# CurveType = 1
# A50 = c(4,4.5)
# A95 = c(6,6.5)
# Pmax = c(1.0,1.0)
# MaturityParams = c(A50, A95, Pmax)
# res=SimulateAgeAtMaturityData(nSamples, CurveType, nSexes, MinAge, MaxAge, MaturityParams)
# AgeCl = res$AgeClasses
# FemObsAgeCl=res$FemObsAgeCl
# MalObsAgeCl=res$MalObsAgeCl
# ObsAgeCl = as.matrix(t(data.frame(FemObsAgeCl=FemObsAgeCl,MalObsAgeCl=MalObsAgeCl)))
# FemObsMatCat=res$FemObsMatCat
# MalObsMatCat=res$MalObsMatCat
# ObsMatCat = as.matrix(t(data.frame(FemObsMatCat=FemObsMatCat,MalObsMatCat=MalObsMatCat)))
# # 2 parameter model
# InitA50 = c(3,3)
# InitA95 = c(5,5)
# params = c(InitA50, InitA95) # without Pmax parameter
# LogisticModType=2
# # 3 parameter model
# # InitA50 = c(4.5,4)
# # InitA95 = c(7,7)
# # InitPmax = c(0.9,0.9)
# # InitPmax_logit = log(InitPmax/(1-InitPmax)) # logit transform
# # params = c(InitPmax_logit, InitA50, InitA95) # with Pmax parameter
# plotages=res$Femplotages # plotting routine assumes the same age range for females and males
# FemPropMat=res$FemPropMat
# MalPropMat=res$MalPropMat
# PropMat = as.matrix(t(data.frame(FemPropMat=FemPropMat,MalPropMat=MalPropMat)))
# FemAgeClSampSize=res$FemAgeClSampSize
# MalAgeClSampSize=res$MalAgeClSampSize
# AgeClSampSize = as.matrix(t(data.frame(FemAgeClSampSize=FemAgeClSampSize,MalAgeClSampSize=MalAgeClSampSize)))
# PlotFittedAgeAtMaturityCurve(params, CurveType, nSexes, ObsAgeCl, ObsMatCat, AgeCl, PropMat, AgeClSampSize, plotages,
#                              xmax=20, xint=2, GraphTitle=NA, xaxis_lab=NA,
#                              yaxis_lab=NA, SampSizelab_cex = NA, ShowSampSizes=FALSE, ShowLegend=FALSE, PlotCLs=FALSE)
}
