% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{PlotFittedGrowthCurve}
\alias{PlotFittedGrowthCurve}
\title{Plot fitted growth curve to fish length-at-age data.}
\usage{
PlotFittedGrowthCurve(
  DataType,
  nSexes,
  GrowthEqn,
  ObsAge,
  ObsLen,
  ObsMeanLen,
  ObsMeanLensd,
  params,
  Ref_ages,
  plotages,
  ymax,
  xmax,
  yint,
  xint,
  GraphTitle,
  xaxis_lab,
  yaxis_lab,
  PlotCLs,
  FittedRes
)
}
\arguments{
\item{DataType}{1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis}

\item{nSexes}{number of sexes}

\item{GrowthEqn}{1=von Bertalanffy, 2=Schnute, 3=Somers seasonal growth curve}

\item{ObsAge}{observed ages}

\item{ObsLen}{observed lengths}

\item{ObsMeanLen}{mean lengths for specified ages (i.e. as estimated from mixture analysis)}

\item{ObsMeanLensd}{sd for mean estimated mean lengths for specified ages (i.e. as estimated from mixture analysis)}

\item{params}{estimated growth parameters}

\item{plotages}{specified ages for plotting}

\item{ymax}{maximum value for y axis}

\item{xmax}{maximum value for x axis}

\item{yint}{interval for y axis}

\item{xint}{interval for x axis}

\item{GraphTitle}{graph title}

\item{xaxis_lab}{label for x axis}

\item{yaxis_lab}{label for y axis}

\item{PlotCLs}{logical True=plot 95 percent confidence limits for curve}
}
\value{
fitted curve on scatter plot with length-at-age data
}
\description{
Plot fitted growth curve to fish length-at-age data.
}
\examples{
# von Bertalanffy growth equation
# simulate data (ignoring mortality and selectivity effects)
GrowthEqn = 1 # von Bertalanffy growth equation
nSexes = 1 # single or combined sex
nSamples = 300
MinAge = 1
MaxAge = 20
AgeStep = 1
Linf = 300
vbK = 0.3
tzero = 0
Growth_params = c(Linf, vbK, tzero)
Ref_ages = NA
Growth_cv = 0.08
Res = SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# fit growth model and plot
DataType=1  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = c(log(400),log(0.3),0) # log(Linf), log(k), tzero
ObsAge = Res$ObsAge
ObsLen = Res$ObsLen
FittedRes=GetvonBertalanffyGrowthResults(params, nSexes, DataType, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLensd=NA)
plotages=seq(0, MaxAge,0.1)
par(mfrow=c(1,1))
PlotFittedGrowthCurve(DataType,  nSexes, GrowthEqn, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLensd=NA,
                      params, Ref_ages, plotages, ymax=600, xmax=NA, yint=100, xint=NA, GraphTitle=NA,
                      xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
DataType=2  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = c(log(400),log(0.3),0) # log(Linf), log(k), tzero
ObsAge = Res$ObsAgeCl
ObsMeanLen = Res$ObsMeanLen
ObsMeanLensd = Res$ObsMeanLensd
FittedRes=GetvonBertalanffyGrowthResults(params, nSexes, DataType, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLensd)
plotages=seq(0, MaxAge,0.1)
PlotFittedGrowthCurve(DataType,  nSexes, GrowthEqn, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLensd,
                      params, Ref_ages, plotages, ymax=600, xmax=NA, yint=100, xint=NA, GraphTitle=NA,
                      xaxis_lab=NA, yaxis_lab=NA, PlotCLs=F, FittedRes)
nSexes = 2 # Separate sexes
nSamples = c(300,200) # if different, order so NAs are last for smaller sample
MinAge = 1
MaxAge = 20
AgeStep = 1
Linf = c(300,250)
vbK = c(0.3,0.3)
tzero = c(0,0)
GrowthEqn = 1
Growth_params = c(Linf, vbK, tzero)
Ref_ages = NA
Growth_cv = c(0.08,0.08)
Res = SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# fit growth model and plot
DataType=1  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
ObsAge = as.matrix(Res$ObsAge)
ObsLen = as.matrix(Res$ObsLen)
params = c(log(c(300,300)),log(c(0.3,0.3)),c(0,0)) # log(Linf), log(k), tzero
FittedRes=GetvonBertalanffyGrowthResults(params, nSexes, DataType, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLensd=NA)
plotages=seq(0, MaxAge,0.1)
par(mfrow=c(1,2))
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLensd=NA,
                      params, Ref_ages, plotages, ymax=600, xmax=NA, yint=100, xint=NA, GraphTitle=NA,
                      xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
DataType=2  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = c(log(c(300,300)),log(c(0.3,0.3)),c(0,0)) # log(Linf), log(k), tzero
ObsAge = Res$ObsAgeCl
FemObsMeanLen = Res$FemObsMeanLen
MalObsMeanLen = Res$MalObsMeanLen
ObsMeanLen = as.matrix(t(data.frame(FemObsMeanLen=FemObsMeanLen,MalObsMeanLen=MalObsMeanLen)))
FemObsMeanLensd = Res$FemObsMeanLensd
MalObsMeanLensd = Res$MalObsMeanLensd
ObsMeanLensd = as.matrix(t(data.frame(FemObsMeanLensd=FemObsMeanLensd,MalObsMeanLensd=MalObsMeanLensd)))
FittedRes=GetvonBertalanffyGrowthResults(params, nSexes, DataType, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLensd)
plotages=seq(0, MaxAge,0.1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLensd,
                      params, Ref_ages, plotages, ymax=600, xmax=NA, yint=100, xint=NA, GraphTitle=NA,
                      xaxis_lab=NA, yaxis_lab=NA, PlotCLs=F, FittedRes)
# Schnute growth equation
# simulate data (ignoring mortality and selectivity effects)
GrowthEqn = 2 # Schnute growth equation
nSexes = 1 # single or combined sex
nSamples = 500
MinAge = 0
MaxAge = 20
AgeStep = 1
ObsAge = runif(nSamples, MinAge, MaxAge)
Ref_ages = c(0.5,15)
Growth_params = c(100,400,0.02,3)
Growth_cv = 0.1
Res=SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# fit growth model and plot
ObsAge=Res$ObsAge
ObsLen=Res$ObsLen
DataType=1  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = c(log(150),log(420),0.05,2)
t1=0.5
t2=15
ObsMeanLen=NA
ObsMeanLensd=NA
FittedRes=GetSchnuteGrowthResults(params, nSexes, DataType, t1, t2, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLensd=NA)
plotages=seq(0, MaxAge,0.1)
par(mfrow=c(1,1))
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn=2, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLensd=NA,
                      params, Ref_ages, plotages, ymax=600, xmax=NA, yint=100, xint=NA, GraphTitle=NA,
                      xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
DataType=2  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = c(log(110),log(420),0.05,2)
t1=0.5
t2=15
ObsAge=Res$ObsAgeCl
ObsMeanLen=Res$ObsMeanLen
ObsMeanLensd=Res$ObsMeanLensd
FittedRes=GetSchnuteGrowthResults(params, nSexes, DataType, t1, t2, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLensd)
plotages=seq(0, MaxAge,0.1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn=2, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLensd,
                      params, Ref_ages, plotages, ymax=600, xmax=NA, yint=100, xint=NA, GraphTitle=NA,
                      xaxis_lab=NA, yaxis_lab=NA, PlotCLs=F, FittedRes)
GrowthEqn = 2 # Schnute equation
nSexes = 2 # separate sexes
nSamples = c(500,500)
MinAge = 0.2
MaxAge = 20
AgeStep = 1
ObsAge = runif(nSamples, MinAge, MaxAge)
Ref_ages = c(0.5,15)
Growth_params = c(100,120,400,450,0.05,0.05,2,2)
Growth_cv = c(0.1,0.1)
Res=SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# Fit Schnute growth curve to length at age data
nSexes=2
ObsAge=as.matrix(Res$ObsAge)
ObsLen=as.matrix(Res$ObsLen)
DataType=1  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = c(log(c(130,125)),log(c(400,450)),0.05,0.05,2,2)
t1=0.5
t2=15
ObsMeanLen=NA
ObsMeanLensd=NA
FittedRes=GetSchnuteGrowthResults(params, nSexes, DataType, t1, t2, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLensd=NA)
par(mfrow=c(1,2))
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn=2, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLensd=NA,
                      params, Ref_ages, plotages, ymax=600, xmax=NA, yint=100, xint=NA, GraphTitle=NA,
                      xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
# Fit Schnute growth curve to mean length at age data
DataType=2  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = c(log(c(100,100)),log(c(400,450)),0.05,0.05,2,2)
ObsAge=Res$ObsAgeCl
FemObsMeanLen = Res$FemObsMeanLen
MalObsMeanLen = Res$MalObsMeanLen
ObsMeanLen = as.matrix(t(data.frame(FemObsMeanLen=FemObsMeanLen,MalObsMeanLen=MalObsMeanLen)))
FemObsMeanLensd = Res$FemObsMeanLensd
MalObsMeanLensd = Res$MalObsMeanLensd
ObsMeanLensd = as.matrix(t(data.frame(FemObsMeanLensd=FemObsMeanLensd,MalObsMeanLensd=MalObsMeanLensd)))
FittedRes=GetSchnuteGrowthResults(params, nSexes, DataType, t1, t2, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLensd)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn=2, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLensd,
                      params, Ref_ages, plotages, ymax=600, xmax=NA, yint=100, xint=NA, GraphTitle=NA,
                      xaxis_lab=NA, yaxis_lab=NA, PlotCLs=F, FittedRes)
# Seasonal growth curve
# simulate data (ignoring mortality and selectivity effects)
GrowthEqn=3 # Seasonal growth curve
nSexes=1 # single or combined sex
nSamples = 1000
MinAge = 0
MaxAge = 3
AgeStep = 1/12
Linf = 150
vbK = 1
tzero = 0
tc = 0.25
C = 0.8
Ref_ages=NA
Growth_params=c(Linf,vbK,tzero,tc,C)
Growth_cv = 0.1
Res=SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# fit growth model and plot
DataType = 1
nSexes=1
ObsAge = Res$ObsAge
ObsLen = Res$ObsLen
ObsMeanLen=NA
ObsMeanLensd=NA
params = c(log(160),log(0.8),0,0.25,0.8) #Linf, vbK, tzero, tc, C
# fit seasonal growth curve to mean length at age data
FittedRes=GetSeasonalGrowthResults(params, nSexes, DataType, ObsAge, ObsLen, ObsMeanLen, ObsMeanLensd)
par(mfrow=c(1,1))
plotages=seq(0, MaxAge,0.1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLensd=NA,
                      params, Ref_ages, plotages, ymax=200, xmax=NA, yint=50, xint=NA, GraphTitle=NA,
                      xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
DataType=2  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = c(log(160),log(0.8),0,0.25,0.8) #Linf, vbK, tzero, tc, C
ObsAge=Res$ObsAgeCl
ObsMeanLen=Res$ObsMeanLen
ObsMeanLensd=Res$ObsMeanLensd
FittedRes=GetSeasonalGrowthResults(params, nSexes, DataType, ObsAge, ObsLen, ObsMeanLen, ObsMeanLensd)
plotages=seq(0, MaxAge,0.1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLensd,
                      params, Ref_ages, plotages, ymax=200, xmax=NA, yint=50, xint=NA, GraphTitle=NA,
                      xaxis_lab=NA, yaxis_lab=NA, PlotCLs=F, FittedRes)
GrowthEqn=3 # Seasonal growth curve
nSexes=2 # separate sexes
nSamples = c(500,500)
MinAge = 0
MaxAge = 3
AgeStep = 1/12
Linf = c(150,160)
vbK = c(1,1)
tzero = c(0,0)
tc = c(0.25,0.25)
C = c(0.8,0.8)
Ref_ages=NA
Growth_params=c(Linf,vbK,tzero,tc,C)
Growth_cv = c(0.1,0.1)
Res=SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# fit growth model and plot
DataType = 1
nSexes=2
ObsAge=as.matrix(Res$ObsAge)
ObsLen=as.matrix(Res$ObsLen)
ObsMeanLen=NA
ObsMeanLensd=NA
params = c(log(c(160,150)),log(c(0.8,0.8)),c(0,0),c(0.25,0.25),c(0.8,0.8)) #Linf, vbK, tzero, tc, C
# fit seasonal growth curve to mean length at age data
FittedRes=GetSeasonalGrowthResults(params, nSexes, DataType, ObsAge, ObsLen, ObsMeanLen, ObsMeanLensd)
par(mfrow=c(1,2))
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLensd=NA,
                      params, Ref_ages, plotages, ymax=200, xmax=NA, yint=50, xint=NA, GraphTitle=NA,
                      xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
DataType=2  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = c(log(c(160,150)),log(c(0.8,0.8)),c(0,0),c(0.25,0.25),c(0.8,0.8)) #Linf, vbK, tzero, tc, C
ObsAge=Res$ObsAgeCl
FemObsMeanLen = Res$FemObsMeanLen
MalObsMeanLen = Res$MalObsMeanLen
ObsMeanLen = as.matrix(t(data.frame(FemObsMeanLen=FemObsMeanLen,MalObsMeanLen=MalObsMeanLen)))
FemObsMeanLensd = Res$FemObsMeanLensd
MalObsMeanLensd = Res$MalObsMeanLensd
ObsMeanLensd = as.matrix(t(data.frame(FemObsMeanLensd=FemObsMeanLensd,MalObsMeanLensd=MalObsMeanLensd)))
FittedRes=GetSeasonalGrowthResults(params, nSexes, DataType, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLensd)
plotages=seq(0, MaxAge,0.1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLensd,
                      params, Ref_ages, plotages, ymax=200, xmax=NA, yint=50, xint=NA, GraphTitle=NA,
                      xaxis_lab=NA, yaxis_lab=NA, PlotCLs=F, FittedRes)
# simulate data allowing for divergent growth
set.seed(123)
GrowthEqn = 4 # von Bertalanffy growth equation with divergent growth
nSexes = 2 # single or combined sex
nSamples = 1000
MinAge = 0
MaxAge = 20
AgeStep = 1
Linf = c(1000,500)
vbK = c(0.2,0.2)
tzero1 = 0
tdiverge = 2
Growth_params = c(Linf, vbK, tzero1, tdiverge)
Ref_ages = NA
Growth_cv = 0.1
Res = SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
x=which(Res$ObsDat$ObsSex==1)
plot(Res$ObsAge[x], Res$ObsLen[x])
y=which(Res$ObsDat$ObsSex==2)
points(Res$ObsAge[y], Res$ObsLen[y],col="blue")
# fit growth model allowing for divergent growth
InitLinf = c(1050,500)
InitvbK = c(0.13,0.13)
Inittzero1 = 0
Inittdiverge = 4
params = c(log(InitLinf), log(InitvbK), Inittzero1, log(Inittdiverge))
ObsAge = Res$ObsDat$ObsAge
ObsLen = Res$ObsDat$ObsLen
ObsSex = Res$ObsDat$ObsSex
FittedRes=GetDivergeGrowthModelResults(params, ObsAge, ObsLen, ObsSex)
FittedRes$ParamEst
DataType=1
plotages = seq(MinAge,MaxAge,0.1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsMeanLen,
                      ObsMeanLensd, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
}
