% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{PlotGSIData}
\alias{PlotGSIData}
\title{Plot mean monthly GSIs}
\usage{
PlotGSIData(
  MatL50,
  FishRepdat,
  ymax,
  yint,
  GraphTitle,
  xaxis_lab = "Mean GSI",
  yaxis_lab = "Month"
)
}
\arguments{
\item{MatL50}{length at first maturity}

\item{FishRepdat}{data frame containing data for individual fish, including month of capture (MM), fish length (FishLen),}

\item{ymax}{maximum value for y axis}

\item{yint}{y axis interval}

\item{GraphTitle}{graph title}

\item{xaxis_lab}{x axis label}

\item{yaxis_lab}{y axis label}
}
\value{
plot of mean month GSIs with 95 percent confidence limits
}
\description{
This function plots mean monthly GSI values and associated 95 percent confidence limits,
used to assess time and duration of spawning
}
\examples{
set.seed(123)
nMnths = 12
MnthSampeSizes = 20
MMSampSize = rep(MnthSampeSizes,nMnths)
TotSampSize = MnthSampeSizes * nMnths
Mnths = 1:nMnths
MeanMMGSI = rep(c(rep(0.1,4),0.25, 0.7, 0.6, 0.3, rep(0.1,4)),MMSampSize)
randMMGSI = rnorm(TotSampSize,MeanMMGSI,0.1)
length(randMMGSI)
MM = rep(Mnths,times=MMSampSize)
FishLen = round(runif(TotSampSize,100,500),0)
FishWt = (0.0000001 * FishLen ^ 3)
GonadWt = (randMMGSI / 100) * FishWt
GSIdat = data.frame(MM=MM, FishLen=FishLen, GonadWt=GonadWt, FishWt=FishWt)
MatL50 = 200
PlotGSIData(MatL50, GSIdat, ymax=NA, yint=NA, GraphTitle="Females",
                                xaxis_lab="Month",yaxis_lab="Mean GSI")
}
