% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{PlotLengthAtAgeData}
\alias{PlotLengthAtAgeData}
\title{Plot fish length-at-age data}
\usage{
PlotLengthAtAgeData(
  ObsAge,
  ObsLen,
  ObsMeanLensd,
  ymax,
  xmax,
  yint,
  xint,
  GraphTitle,
  xaxis_lab,
  yaxis_lab,
  PlotCLs
)
}
\arguments{
\item{ObsAge}{observed ages}

\item{ObsLen}{observed lengths}

\item{ObsMeanLensd}{sd for mean estimated mean lengths for specified ages (i.e. as estimated from mixture analysis)}

\item{ymax}{maximum value for y axis}

\item{xmax}{maximum value for x axis}

\item{yint}{interval for y axis}

\item{xint}{interval for x axis}

\item{GraphTitle}{graph title}

\item{xaxis_lab}{label for x axis}

\item{yaxis_lab}{label for y axis}
}
\value{
scatter plot of length-at-age data
}
\description{
Plots either raw data (PlotCLs=FALSE) or
mean lengths at age with specified standard deviations for each mean length
(PlotCLs=TRUE)
}
\examples{
# plot mean lengths at age with approx 95\% CLs given sd at length
nSamples = 300
ObsAge = 1:20
MeanLen = 300 * (1 - exp(-0.3 * (ObsAge - 0)))
ObsLen = MeanLen
ObsMeanLensd = MeanLen*0.1
PlotLengthAtAgeData(ObsAge, ObsLen, ObsMeanLensd, ymax=NA, xmax=NA, yint=NA, xint=NA,
                   GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=TRUE)
# plot raw lengths at age
set.seed(123)
nSamples = 300
ObsAge = runif(nSamples, 1, 20)
MeanLen = 300 * (1 - exp(-0.3 * (ObsAge - 0)))
ObsLen = rnorm(nSamples, MeanLen, 0.1 * MeanLen)
PlotLengthAtAgeData(ObsAge, ObsLen, ObsMeanLensd=NA, ymax=NA, xmax=NA, yint=NA, xint=NA,
                   GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=TRUE)
}
