% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{PlotLengthAtMaturityData}
\alias{PlotLengthAtMaturityData}
\title{Plot observed fish maturity-at-length data}
\usage{
PlotLengthAtMaturityData(
  nSexes,
  midpt,
  PropMat,
  LenClSampSize,
  xmax,
  xint,
  GraphTitle,
  SampSizelab_cex,
  xaxis_lab,
  yaxis_lab
)
}
\arguments{
\item{nSexes}{number of sexes}

\item{midpt}{length class mid points}

\item{PropMat}{proportion mature in length classes}

\item{LenClSampSize}{length class sample sizes}

\item{xmax}{x axis maximum}

\item{xint}{x axis interval}

\item{GraphTitle}{graph title}

\item{SampSizelab_cex}{sample size graph label size}

\item{xaxis_lab}{x axis label}

\item{yaxis_lab}{y axis label}
}
\value{
scatter plot of maturity-at-length data
}
\description{
This function produces a scatter plot plot of fish length-at-maturity data
}
\examples{
# generate some synthetic maturity data
set.seed(123)
MaxAge = 20
MinLen = 0
MaxLen = 400
LenInc = 20
# single sex/combined sexes
nSexes = 1
nSamples = 300
CurveType = 1 # 1 = symmetric logistic, 2 = asymmetric logistic
Linf = 300
vbK = 0.3
tzero = 0
CVSizeAtAge = 0.1
GrowthParams = c(Linf, vbK, tzero, CVSizeAtAge)
L50 = 220
L95 = 270
Pmax = 1.0
MaturityParams = c(L50, L95, Pmax)
# nSexes = 2
# nSamples = c(300,300)
# CurveType = 1 # 1 = symmetric logistic, 2 = asymmetric logistic
# Linf = c(300,350)
# vbK = c(0.3,0.3)
# tzero = c(0,0)
# CVSizeAtAge = c(0.1,0.1)
# GrowthParams = c(Linf, vbK, tzero, CVSizeAtAge)
# L50 = c(220,250)
# L95 = c(270,290)
# Pmax = c(1.0,1.0)
# MaturityParams = c(L50, L95, Pmax)
res=SimulateLengthAtMaturityData(nSamples, CurveType, nSexes, MaxAge, MinLen, MaxLen, LenInc, GrowthParams, MaturityParams)
midpt = res$midpt
# single sex
PropMat = res$PropMat
LenClSampSize=res$LenClSampSize
# separate sexes
# PropMat = t(data.frame(FemPropMat=res$FemPropMat,MalPropMat=res$MalPropMat))
# LenClSampSize = t(data.frame(FemLenClSampSize=res$FemLenClSampSize,MalLenClSampSize=res$MalLenClSampSize))
PlotLengthAtMaturityData(nSexes=1, midpt, PropMat, LenClSampSize, xmax=NA, xint=NA,
                         GraphTitle=NA, SampSizelab_cex=NA, xaxis_lab=NA, yaxis_lab=NA)
}
