% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{PlotMeanMonthyStGonadWts}
\alias{PlotMeanMonthyStGonadWts}
\title{Plot mean monthly standardised gonad weights}
\usage{
PlotMeanMonthyStGonadWts(
  MatL50,
  SpecFishLength,
  FishRepdat,
  ymax,
  yint,
  GraphTitle,
  xaxis_lab,
  yaxis_lab,
  SampSizelabPosAdj,
  SampSizelab_cex
)
}
\arguments{
\item{MatL50}{length at first maturity}

\item{SpecFishLength}{specified fish lengths to which gonad weights are to be standardised}

\item{FishRepdat}{data frame containing data for individual fish, including month of capture (Mnth), fish length (FishLen),}

\item{ymax}{maximum value for y axis}

\item{yint}{y axis interval}

\item{GraphTitle}{graph title}

\item{xaxis_lab}{x axis label}

\item{yaxis_lab}{y axis label}

\item{SampSizelabPosAdj}{position of monthly sample sizes on plot above upper 95 percent CLs}

\item{SampSizelab_cex}{size of monthly sample size labels}
}
\value{
plot of mean month GSIs with 95 percent confidence limits
}
\description{
This function plots mean monthly standardised gonad weights (for specified fish length value) and associated
95 percent confidence limits, used to assess time and duration of spawning
}
\examples{
# Generate synthetic length composition data from a simple population model
library(L3Assess)
set.seed(123)
SampleSize=5000 # sample size for retained catches (and same number for released fish, if an MLL is specified)
MaxAge = 30
TimeStep = 1 # model timestep (e.g. 1 = annual, 1/12 = monthly)
NatMort = 4.22/MaxAge
FishMort = 0.2
MaxLen = 1200
LenInc = 1
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified
SelectivityType=2 # 1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve
SelectivityVec = NA # selectivity vector
SelParams = c(300, 50) # L50, L95-L50 for gear selectivity
RetenParams = c(500, 50) # L50, L95-L50 for retention
DiscMort = 0 # proportion of fish that die due to natural mortality
GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
Linf = 800
vbK = 0.2
CVSizeAtAge = 0.08
GrowthParams = c(Linf, vbK)
RefnceAges = NA
Res=SimLenAndAgeFreqData(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                         SelParams, RetenParams, SelectivityVec, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
# Simulate length-frequency data from Dirchlet multinomial distribution.
# install.packages("dirmult")
library(dirmult)
set.seed(123)
nSampEvents = 5 # number of sampling events per month
nFishPerSampEvent = 5 # number of fish caught per sampling event
theta_val = 0.9 # level of autocorrelation of lengths within sampling events
midpt = Res$midpt
ExpPropAtLen = Res$ModelDiag$ExpRetCatchPropAtLen
# Generate fish lengths (taken as midpts of length classes)
FishLen <- NA
for (mm in 1:12) {
  res=SimLenFreqDat_DirMultDistn(nSampEvents, nFishPerSampEvent, theta_val, midpt, ExpPropAtLen)
  if (mm == 1) {
    FishLen <- rep(res$midpt, res$simLenFreq)
  } else {
    tempFishLen <- rep(res$midpt, res$simLenFreq)
    FishLen <- c(FishLen,tempFishLen)
  }
}
FishWt <- 0.00002 * FishLen ^ 3
# specify monthly true trend in standardised gonad weight
MnthEff <- c(0.22,0.25,0.26,0.28,0.35,0.54,0.45,0.33,0.21,0.18,0.25,0.22)
Mnth <- sort(rep(1:12,nSampEvents*nFishPerSampEvent))
a1 <- 0.25 # fish length effect
GonadWt <- rep(0,length(FishLen))
for (i in 1:length(FishLen)) {
  tempMnth <- Mnth[i]
  GonadWt[i] <-  MnthEff[tempMnth] * (a1 * FishLen[i]) + rnorm(1,0,10)
}
GonadWt[which(GonadWt<0.01)] <- 0.01
FishRepdat <- data.frame(GonadWt, FishLen, FishWt, Mnth)
# Create scatter plots of the dependent variable by the covariates
# par(mfrow = c(3, 2)) # Set the layout of the plots
# plot(FishWt ~ FishLen, data = FishRepdat, main = "Fish weight by Fish length")
# boxplot(FishLen ~ Mnth, data = FishRepdat, main = "FishLen by Month")
# boxplot(GonadWt ~ Mnth, data = FishRepdat, main = "GonadWt by Month")
# hist(FishLen)
# hist(FishWt)
PlotMeanMonthyStGonadWts(MatL50=200, SpecFishLength=600, FishRepdat, ymax=100, yint=20, GraphTitle="Females",
                         xaxis_lab="Month", yaxis_lab="Stand. gonad wt.", SampSizelabPosAdj=2, SampSizelab_cex=0.8)
}
