% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{PlotMixtureDistnResults}
\alias{PlotMixtureDistnResults}
\title{Plot fitted mixture model to size frequency data.}
\usage{
PlotMixtureDistnResults(
  params,
  DistnType,
  Cohorts,
  ObsFreq,
  MinSize,
  MaxSize,
  SizeInt,
  ymax,
  xmax,
  yint,
  xint,
  GraphTitle,
  xaxis_lab,
  yaxis_lab,
  set.plot.par,
  PlotHist,
  PlotCLs
)
}
\arguments{
\item{params}{c(Mean1, sd1) normal distn: 1 mode, or
c(Shape1, Rate1) gamma distn: 1 mode, or
c(Mean1, Mean2, sd1, PropZero) normal distn: 2, common sd, or
c(Shape1, Shape2, Rate1, PropZero) gamma distn: 2, common sd, or
c(Mean1, Mean2, sd1, sd2, PropZero) normal distn: 2, separate sds, or
c(Shape1, Shape2, Rate1, Rate2, PropZero) gamma distn: 2 modes, separate sds, or
c(Mean1, Mean2, Mean3, sd1, sd2, sd3, PropZero, PropOne) normal distn: 3 modes, separate sds, or
c(Shape1, Shape2, Shape3, Rate1, Rate2, Rate3, PropZero, PropOne) gamma distn: 3 modes, separate sds}

\item{DistnType}{1=normal, 2=gamma}

\item{Cohorts}{1, 2 or 3}

\item{ObsFreq}{observed frequency data}

\item{MinSize}{minimum size}

\item{MaxSize}{maximum size}

\item{SizeInt}{size class interval}

\item{ymax}{y axis maximum}

\item{xmax}{x axis maximum}

\item{yint}{interval for y axis}

\item{xint}{interval for x axis}

\item{GraphTitle}{title for graph}

\item{xaxis_lab}{label for x axis}

\item{yaxis_lab}{label for y axis}

\item{set.plot.par}{logical, true sets program default}

\item{PlotHist}{logical, true is histogram, false is scatter plot}

\item{PlotCLs}{logical, true plots 95 percent confidence limits}
}
\value{
fitted mixture distribution curves either scatter plot or histogram
}
\description{
Plot fitted mixture model to size frequency data.
}
\examples{
# Simulate data with 1 mode, specifying normal distribution
set.seed(123)
MinSize = 0
MaxSize = 60
SizeInt = 1
SampSize = 1000
Cohorts=1
Mean1 = 20
sd1 = 5
ObsSize = round(rnorm(SampSize, Mean1, sd1),0)
HistData=hist(ObsSize, breaks=seq(0,MaxSize,SizeInt), right=FALSE, plot=FALSE)
ObsFreq = HistData$counts
# Specify starting parameter values for Mean1 and sd1, and return associated negative log-likelihood
DistnType = 1 # 1=normal, 2=gamma
params = log(c(30, 5)) # normal
PlotMixtureDistnResults(params, DistnType, Cohorts, ObsFreq, MinSize, MaxSize, SizeInt, ymax=100, xmax=50, yint=20, xint=10,
GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, set.plot.par=TRUE, PlotHist=TRUE, PlotCLs=FALSE)
# Simulate data with 1 mode, specifying gamma distribution
set.seed(123)
MinSize = 0
MaxSize = 60
SizeInt = 1
SampSize = 1000
Cohorts=1
Shape1 = 20
Rate1 = 1
ObsSize = round(rgamma(SampSize, Shape1, Rate1),0)
HistData=hist(ObsSize, breaks=seq(0,MaxSize,SizeInt), right=FALSE, plot=FALSE)
ObsFreq = HistData$counts
DistnType = 2 # 1=normal, 2=gamma
params = log(c(20, 1)) # gamma
PlotMixtureDistnResults(params, DistnType, Cohorts, ObsFreq, MinSize, MaxSize, SizeInt, ymax=100, xmax=50, yint=20, xint=10,
GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, set.plot.par=TRUE, PlotHist=TRUE, PlotCLs=FALSE)
# Simulate data with 2 modes, specifying normal distribution
set.seed(123)
MinSize = 0
MaxSize = 60
SizeInt = 1
SampSize = 1000
# normal
Mean1 = 20
Mean2 = 35
sd1 = 5
sd2 = 5
PropZero = 0.7
ObsSize1 = round(rnorm(PropZero*SampSize, Mean1, sd1),0)
ObsSize2 = round(rnorm((1-PropZero)*SampSize, Mean2, sd2),0)
ObsSize = c(ObsSize1,ObsSize2)
HistData=hist(ObsSize, breaks=seq(0,MaxSize,SizeInt), right=FALSE, plot=FALSE)
ObsFreq = HistData$counts
params = c(log(c(15, 40, 5, 5)), 0.5) # separate sd's for the 2 cohorts
Cohorts=2
DistnType = 1 # 1=normal, 2=gamma
PlotMixtureDistnResults(params, DistnType, Cohorts, ObsFreq, MinSize, MaxSize, SizeInt, ymax=100, xmax=50, yint=20, xint=10,
GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, set.plot.par=TRUE, PlotHist=TRUE, PlotCLs=FALSE)
# Simulate data with 2 modes, specifying gamma distribution
set.seed(123)
MinSize = 0
MaxSize = 60
SizeInt = 1
SampSize = 1000
Shape1 = 20
Shape2 = 35
Rate1 = 1.5
Rate2 = 1
PropZero = 0.7
ObsSize1 = round(rgamma(PropZero*SampSize, Shape1, Rate1),0)
ObsSize2 = round(rgamma((1-PropZero)*SampSize, Shape2, Rate2),0)
ObsSize = c(ObsSize1,ObsSize2)
HistData=hist(ObsSize, breaks=seq(0,MaxSize,SizeInt), right=FALSE, plot=TRUE)
ObsFreq = HistData$counts
params = c(log(c(15, 40, 1, 1)), 0.5) # separate sd's for the 2 cohorts
Cohorts=2
DistnType = 2 # 1=normal, 2=gamma
PlotMixtureDistnResults(params, DistnType, Cohorts, ObsFreq, MinSize, MaxSize, SizeInt, ymax=100, xmax=50, yint=20, xint=10,
GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, set.plot.par=TRUE, PlotHist=TRUE, PlotCLs=FALSE)
}
