% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{PlotMonthlyGonadStageProps}
\alias{PlotMonthlyGonadStageProps}
\title{Plot of calculated monthly proportions of fish gonad stages}
\usage{
PlotMonthlyGonadStageProps(MatL50, FishRepdat)
}
\arguments{
\item{MatL50}{length at first maturity}

\item{FishRepdat}{data frame containing data for individual fish, including month of capture (MM), and fish length (FishLen)}
}
\value{
plot of monthly proportions of fish gonad stages
}
\description{
Plot of calculated monthly proportions of fish gonad stages, used to assess time and duration of spawning
}
