% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{PlotSpawningDurationData}
\alias{PlotSpawningDurationData}
\title{Plot daily proportions of spawning capable fish for specified length categories}
\usage{
PlotSpawningDurationData(ObsSpawnDat, lbnds, ubnds, params)
}
\arguments{
\item{ObsSpawnDat}{list including observed fish lengths (FishLen), observed months of capture (MM)
and observed maturity status categories (ObsMatStatus, 0=immature, 1=mature)}

\item{lbnds}{lower bounds for length classes to be plotted}

\item{ubnds}{upper bounds for length classes to be plotted}

\item{params}{specified (or estimated), transformed parameter values, for the spawning duration model (including
lnL50, lnslope, logitPkSpawn, lnkappa, lnkappa2, lnslope1, lnslope2)}
}
\value{
plots
}
\description{
This function plots calculated values for the daily proportions of spawning capable fish
for specified length categories and estimated daily proportions from spawning duration model,
(i.e. asymmetric double logistic model with variable height), if parameters are specified
}
\examples{
# Simulate spawning proportion data
# First, simulate length and age data
set.seed(123)
GrowthEqn=1 # von Bertalanffy
nSamples = 5000
nSexes = 1
MinAge = 1
MaxAge = 20
AgeStep = 1
Ref_ages = NA
Linf = 1000
vbK = 0.1
tzero = 0
Growth_params = c(Linf,vbK,tzero)
Growth_cv = 0.08
# Spawning duration model parameters
lnL50 = log(300)
lnslope= log(0.015)
logitPkSpawn = log(0.6/(1-0.6))
lnkappa = log(0.0002)
lnkappa2 = log(0.0002)
lnslope1 = log(30)
lnslope2 = log(30)
params = c(lnL50,lnslope,logitPkSpawn,lnkappa,lnkappa2,lnslope1,lnslope2)
Res=SimulateSpawningDurationData(params, GrowthEqn, nSamples, nSexes, MinAge, MaxAge,
                                 AgeStep, Ref_ages, Growth_params, Growth_cv)
ObsMatStatus=Res$ObsMatStatus
FishLen=Res$FishLen
MM=Res$MM
ObsSpawnDat = data.frame(MM=MM,FishLen=FishLen,ObsMatStatus=ObsMatStatus)
# params = c(lnL50,lnslope,logitPkSpawn,lnkappa,lnkappa2,lnslope1,lnslope2)
params = NA
CalcNLL_SpDurMod(params)
params = NA
lbnds = seq(300,900,100)
ubnds = lbnds + 100
PlotSpawningDurationData(ObsSpawnDat, lbnds, ubnds, params)
}
