% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{PlotSpawningDurationModelResults}
\alias{PlotSpawningDurationModelResults}
\title{Plot results from spawning duration model}
\usage{
PlotSpawningDurationModelResults(
  DecDay,
  ObsSpawnDat,
  MinLen,
  MaxLen,
  LenInc,
  params,
  nsims,
  FittedRes
)
}
\arguments{
\item{DecDay}{decimal days of fish capture during the year}

\item{ObsSpawnDat}{list including observed fish lengths (FishLen), observed months of capture (MM)
and observed maturity status categories (ObsMatStatus, 0=immature, 1=mature)}

\item{MinLen}{Minimum fish length (for determining lengths to output expected model curves)}

\item{MaxLen}{maximum length of fish (for plotting data for length classes)}

\item{LenInc}{length increments of length classes, for plotting}

\item{params}{initial (transformed) values of model parameters, for the spawning duration model (including
lnL50, lnslope, logitPkSpawn, lnkappa, lnkappa2, lnslope1, lnslope2)}

\item{nsims}{number of resampled parameter values, used to calculate confidence intervals}

\item{FittedRes}{Outputs of fitted spawning duration model using the function GetSpawningDurationModelResults}

\item{lbnds}{lower bounds for length classes to be plotted}

\item{ubnds}{upper bounds for length classes to be plotted}
}
\value{
plots associated with fit of spawning duration model
}
\description{
This function plots observed for the daily proportions of spawning capable fish for specified length categories,
and estimated daily proportions from spawning duration model (with associated 95 percent confidence limits),
(i.e. asymmetric double logistic model with variable height), and estimated relationships between proportion spawning
vs fish length at the estimated time of peak spawning, and of estimated spawning duration vs fish length
}
\examples{
# Simulate spawning proportion data
# First, simulate length and age data
set.seed(123)
GrowthEqn=1 # von Bertalanffy
nSamples = 5000
nSexes = 1
MinAge = 1
MaxAge = 20
AgeStep = 1
Ref_ages = NA
Linf = 1000
vbK = 0.1
tzero = 0
Growth_params = c(Linf,vbK,tzero)
Growth_cv = 0.08
# Spawning duration model parameters
lnL50 = log(300)
lnslope= log(0.015)
logitPkSpawn = log(0.6/(1-0.6))
lnkappa = log(0.0002)
lnkappa2 = log(0.0002)
lnslope1 = log(30)
lnslope2 = log(30)
params = c(lnL50,lnslope,logitPkSpawn,lnkappa,lnkappa2,lnslope1,lnslope2)
Res=SimulateSpawningDurationData(params, GrowthEqn, nSamples, nSexes, MinAge, MaxAge,
                                 AgeStep, Ref_ages, Growth_params, Growth_cv)
ObsMatStatus=Res$ObsMatStatus
FishLen=Res$FishLen
MM=Res$MM
DecDay=Res$DecDay
# Spawning duration model parameters - initial values
lnL50 = log(350)
lnslope= log(0.02)
logitPkSpawn = log(0.7/(1-0.7))
lnkappa = log(0.00015)
lnkappa2 = log(0.00035)
lnslope1 = log(25)
lnslope2 = log(35)
params = c(lnL50,lnslope,logitPkSpawn,lnkappa,lnkappa2,lnslope1,lnslope2)
nsims=20
MinLen=0
MaxLen=1100
LenInc=100
FittedRes=GetSpawningDurationModelResults(params, FishLen, DecDay, ObsMatStatus, MaxLen, LenInc, nsims)
PlotSpawningDurationModelResults(DecDay, ObsSpawnDat, MinLen, MaxLen, LenInc, params, nsims, FittedRes)
}
