% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{PlotWeightLengthRel_NormalSpace}
\alias{PlotWeightLengthRel_NormalSpace}
\title{Plot weight length relationship over data in normal space}
\usage{
PlotWeightLengthRel_NormalSpace(
  FishLen,
  FishWt,
  xmin,
  ymin,
  xmax,
  ymax,
  xint,
  yint,
  GraphTitle,
  xaxis_lab,
  yaxis_lab,
  PlotCLs
)
}
\arguments{
\item{FishLen}{vector of fish lengths}

\item{FishWt}{vector of fish weights}

\item{xmin}{minimum value for x axis}

\item{ymin}{minimum value for y axis}

\item{xmax}{maximum value for x axis}

\item{ymax}{maximum value for y axis}

\item{xint}{interval for x axis}

\item{yint}{interval for y axis}

\item{GraphTitle}{title for graph}

\item{xaxis_lab}{x axis label}

\item{yaxis_lab}{y axis label}

\item{PlotCLs}{logical (set to TRUE to plot 95 percent confidence and prediction limits)}
}
\value{
fitted curve on scatter plot with weight-length data
}
\description{
Plot weight length relationship over data in normal space
}
\examples{
# generate synthetic data
nsamples = 100
minlen = 20
maxlen = 500
lenwt_a = 0.00002
lenwt_b = 3
lenwt_sd = 0.1
Res = SimulateWeightLengthData(nsamples, minlen, maxlen,
                               lenwt_a, lenwt_b, lenwt_sd)
FishLen=Res$FishLen
FishWt=Res$FishWt/1000 # kg
# Fit weight-length relationship and plot results (in normal space)
PlotWeightLengthRel_NormalSpace(FishLen, FishWt, xmin=NA, ymin=NA, xmax=NA, ymax=NA, xint=NA, yint=NA,
GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T)
}
