% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{SchnuteGrowthfunction}
\alias{SchnuteGrowthfunction}
\title{Calculate estimated length at age from Schnute growth curve}
\usage{
SchnuteGrowthfunction(Age, t1, t2, y1, y2, a, b)
}
\arguments{
\item{Age}{specified age}

\item{t1}{first reference age}

\item{t2}{second reference age}

\item{y1}{length at t1}

\item{y2}{length at t2}

\item{a}{growth curve parameter}
}
\value{
Estimated length at a specified age
}
\description{
Calculates estimated length at age from Schnute growth curve, given the age, growth parameters and
reference ages
}
\details{
Function requires two reference ages, t1 and t2 (stored in memory in R).
}
\examples{
t1=0.5; t2=10
y1=50; y2=250; a=0; b=3
Age=2
SchnuteGrowthfunction(Age, t1, t2, y1, y2, a, b)
}
