% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{SimulateLengthAtAgeData}
\alias{SimulateLengthAtAgeData}
\title{Simulate fish length at age data from von Bertalanffy growth curve}
\usage{
SimulateLengthAtAgeData(
  GrowthEqn,
  nSamples,
  nSexes,
  MinAge,
  MaxAge,
  AgeStep,
  Ref_ages,
  Growth_params,
  Growth_cv
)
}
\arguments{
\item{GrowthEqn}{1=von Bertalanffy, 2=Schnute, 3=Somers seasonal curve}

\item{nSamples}{number of samples}

\item{nSexes}{number of sexes 1=single or combined sex, 2=separate sexes}

\item{MinAge}{minimum age}

\item{MaxAge}{maximum age}

\item{AgeStep}{age interval}

\item{Linf}{asymptotic length}

\item{vbK}{von Bertalanffy growth coefficient}

\item{tzero}{hypothetical length at age zero}

\item{Growth_sd}{standard deviation of lengths at age}
}
\value{
lengths for individual fish (ObsLen), age classes in sample (ObsAgeCl),
mean length for each age class for females and males or combined sexes (ObsMeanLen,
FemObsMeanLen, MalObsMeanLen), standard deviations for mean
lengths at age for females and males or combined sexes (ObsMeanLensd, FemObsMeanLensd,
MalObsMeanLensd), lower and upper 95 percent confidence limits for
mean lengths at age for females and males or combined sexes (ObsMeanLen_lw95CL, ObsMeanLen_hi95CL,
FemObsMeanLen_lw95CL, MalObsMeanLen_lw95CL, FemObsMeanLen_hi95CL, MalObsMeanLen_hi95CL)
}
\description{
Simulates fish length at age data for individual fish, and mean length at age data
}
\examples{
set.seed(123)
# 1 sex
nSexes = 1
nSamples = 300
MinAge = 1
MaxAge = 20
AgeStep = 1
Linf = 300
vbK = 0.3
tzero = 0
Growth_cv = 0.08
Res = SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# plot lengths at age
plot(Res$ObsAge, Res$ObsLen, ylim=c(0,600))
# plot mean length at age data
ObsAgeCl=Res$ObsAgeCl
ObsMeanLen=Res$ObsMeanLen
ObsMeanLen_lw95CL=Res$ObsMeanLen_lw95CL
ObsMeanLen_hi95CL=Res$ObsMeanLen_hi95CL
plot(ObsAgeCl, ObsMeanLen, ylim=c(0,500))
arrows(ObsAgeCl, ObsMeanLen_lw95CL, ObsAgeCl, ObsMeanLen_hi95CL,
       code=3, angle=90,length=0.02, col='black')
# # 2 sexes
# nSexes = 2
# nSamples = c(300,300)
# MinAge = 1
# MaxAge = 20
# Linf = c(300,250)
# vbK = c(0.3,0.3)
# tzero = c(0,0)
# Growth_cv = c(0.08,0.08)
# Res = SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# # plot lengths at age
# for (i in 1:nSexes) {
#   ObsAge=as.vector(unlist(Res$ObsAge[i,]))
#   ObsLen=as.vector(unlist(Res$ObsLen[i,]))
#   if (i==1) {
#     plot(ObsAge, ObsLen, ylim=c(0,500))
#   } else {
#     points(ObsAge, ObsLen, col="blue")
#   }
# }
# # plot mean lengths at age
# ObsAgeCl=Res$ObsAgeCl
# ObsMeanLen=Res$FemObsMeanLen
# ObsMeanLen_lw95CL=Res$FemObsMeanLen_lw95CL
# ObsMeanLen_hi95CL=Res$FemObsMeanLen_hi95CL
# plot(ObsAgeCl, ObsMeanLen, ylim=c(0,500))
# arrows(ObsAgeCl, ObsMeanLen_lw95CL, ObsAgeCl, ObsMeanLen_hi95CL,
#        code=3, angle=90,length=0.02, col='black')
# ObsMeanLen=Res$MalObsMeanLen
# ObsMeanLen_lw95CL=Res$MalObsMeanLen_lw95CL
# ObsMeanLen_hi95CL=Res$MalObsMeanLen_hi95CL
# points(ObsAgeCl, ObsMeanLen, col="blue")
# arrows(ObsAgeCl, ObsMeanLen_lw95CL, ObsAgeCl, ObsMeanLen_hi95CL,
#        code=3, angle=90,length=0.02, col='blue')
}
