% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{SimulateSpawningDurationData}
\alias{SimulateSpawningDurationData}
\title{Simulated data with daily proportions of spawning capable fish for a species, based on using spawning duration model}
\usage{
SimulateSpawningDurationData(
  params,
  GrowthEqn,
  nSamples,
  nSexes,
  MinAge,
  MaxAge,
  AgeStep,
  Ref_ages,
  Growth_params,
  Growth_cv
)
}
\arguments{
\item{params}{specified (or estimated), transformed parameter values, for the spawning duration model (including
lnL50, lnslope, logitPkSpawn, lnkappa, lnkappa2, lnslope1, lnslope2)}

\item{GrowthEqn}{1=von Bertalanffy, 2=Schnute, 3=Somers seasonal curve}

\item{nSamples}{number of samples}

\item{nSexes}{number of sexes 1=single or combined sex, 2=separate sexes}

\item{MinAge}{minimum age}

\item{MaxAge}{maximum age}

\item{AgeStep}{age interval}

\item{Linf}{asymptotic length}

\item{vbK}{von Bertalanffy growth coefficient}

\item{tzero}{hypothetical length at age zero}

\item{Growth_sd}{standard deviation of lengths at age}
}
\value{
returns a list object containing input data for dates of fish capture in decimal days (DecDay),
month of capture (MM), corresponding fish lengths (FishLen), expected probability of each fish spawning
according to its date of capture and size (ExpSpawnProbs), and randomly-generated maturity status (ObsMatStatus, 0=immature,
1=mature)
}
\description{
This function simulates data containing daily proportions of spawning capable fish for a specified with
specified growth characteristics, and spawning characteristics according to values of specified parameters
for the spawning duration model (i.e. asymmetric double logistic model with variable height)
}
\examples{
# Simulate spawning proportion data
# First, simulate length and age data
set.seed(123)
GrowthEqn=1 # von Bertalanffy
nSamples = 5000
nSexes = 1
MinAge = 1
MaxAge = 20
AgeStep = 1
Ref_ages = NA
Linf = 1000
vbK = 0.1
tzero = 0
Growth_params = c(Linf,vbK,tzero)
Growth_cv = 0.08
# Spawning duration model parameters
lnL50 = log(300)
lnslope= log(0.015)
logitPkSpawn = log(0.6/(1-0.6))
lnkappa = log(0.0002)
lnkappa2 = log(0.0002)
lnslope1 = log(30)
lnslope2 = log(30)
params = c(lnL50,lnslope,logitPkSpawn,lnkappa,lnkappa2,lnslope1,lnslope2)
Res=SimulateSpawningDurationData(params, GrowthEqn, nSamples, nSexes, MinAge, MaxAge,
                                 AgeStep, Ref_ages, Growth_params, Growth_cv)
}
