% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{SimulateTimeVaryingLengthAtAgeData}
\alias{SimulateTimeVaryingLengthAtAgeData}
\title{Simulate multiple years of length at age data with time-varying growth}
\usage{
SimulateTimeVaryingLengthAtAgeData(
  Linf,
  vbK,
  tzero,
  CV,
  nLinfVals,
  nvbKVals,
  StartYear,
  EndYear,
  minIntAge,
  MaxIntAge,
  nSexes
)
}
\arguments{
\item{Linf}{asymptotic length (by sex, and year, if multiple years)}

\item{vbK}{growth coefficient (by sex, and year, if multiple years)}

\item{tzero}{combined sex, single value}

\item{CV}{coefficient of variation, constant for mean length at age}

\item{nLinfVals}{number of consecutive years of Linf params}

\item{nvbKVals}{number of consecutive years of vbK params}

\item{StartYear}{first year of length-at-age data}

\item{EndYear}{last year of length-at-age data}

\item{minIntAge}{minimum integer age}

\item{MaxIntAge}{maximum integer age}

\item{nSexes}{number of sexes}
}
\value{
results (RandLenAtAge, EstLenAtAge, EstLenAtIntAge, nYrs, nIntAges, nObs, LenAtAgeDat)
}
\description{
This function simulates multiple connsecutive years of length-at-age data
with time-varying growth, based on year specific von Bertalanffy growth
parameters (Linf and/or vbK). Data are simulated assuming constant mortality
and selectivity.
}
\examples{
library(L3Assess)
library(RTMB)
set.seed(123)
MinAge = 1
MaxAge = 40
Ages = MinAge:MaxAge
NatMort = 0.11
FMort = 0.11
ZMort = FMort + NatMort
SelA50 = c(7,6)
SelA95 = c(8,7)
SampleSize = 300 # required sample size. For 2 sex model, same sample size generated for each sex.
Res=SimAgeFreqData_EqMod(SampleSize, MinAge, MaxAge, SelA50, SelA95, NatMort, FMort)
ObsAgeFreq = Res$CatchSample
par(mfrow=c(1,1),mar=c(4,3,2,2))
plot(MinAge:MaxAge, ObsAgeFreq[1,],col="red",'o')
lines(MinAge:MaxAge, ObsAgeFreq[2,],col="blue")
# Specify information required to generate length-at-age data with
# time-varying growth
StartYear = 2000
EndYear = 2020
nYrs = EndYear - StartYear + 1
minIntAge = 0
MaxIntAge = 40
nSexes = 2
# Specify true growth parameter values. Linf and/or vbK can be year-specific.
nLinfVals = 1
nvbKVals = nYrs
Linf_Fem = rep(1000,nLinfVals)
Linf_Mal = rep(1100,nLinfVals)
Linf = data.frame(data.frame(Linf_Fem=Linf_Fem,Linf_Mal=Linf_Mal))
vbK_Fem = seq(0.1,0.2,0.1/nYrs)
vbK_Mal = seq(0.1,0.2,0.1/nYrs)
vbK = data.frame(data.frame(vbK_Fem=vbK_Fem,vbK_Mal=vbK_Mal))
tzero = 0 # Specify common tzero for both sexes to simplify
CV = 0.1
Res = SimulateTimeVaryingLengthAtAgeData(Linf, vbK, tzero, CV, nLinfVals, nvbKVals, StartYear, EndYear, minIntAge, MaxIntAge, nSexes)
}
