% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{SimulateWeightLengthData}
\alias{SimulateWeightLengthData}
\title{Simulate weight-length data}
\usage{
SimulateWeightLengthData(nsamples, minlen, maxlen, lenwt_a, lenwt_b, lenwt_sd)
}
\arguments{
\item{nsamples}{number of required samples}

\item{minlen}{minimum length}

\item{maxlen}{maximum length}

\item{lenwt_a}{weight length equation parameter}

\item{lenwt_b}{weight length equation parameter}

\item{lenwt_sd}{standard deviation of data (in log space)}
}
\value{
randomly-generated fish lengths (FishLen) and fish weights (FishWt)
}
\description{
Function simulates weight-length data for individual fish
}
\examples{
# generate synthetic data
set.seed(123)
nsamples = 100
minlen = 20
maxlen = 500
lenwt_a = 0.00002
lenwt_b = 3
lenwt_sd = 0.1
Res = SimulateWeightLengthData(nsamples, minlen, maxlen,
                               lenwt_a, lenwt_b, lenwt_sd)
Res
}
