% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{x_is_gt_bound}
\alias{x_is_gt_bound}
\title{Logistic function (for lower bound) of penalty function for smooth upper and lower bounds (attributed to Prof. Norm Hall)}
\usage{
x_is_gt_bound(x, bound, slope)
}
\arguments{
\item{x}{value to be check for penalty constraint}
}
\value{
Returns 0 if x << bound and 1 if x >> bound
}
\description{
Returns 0 if x << bound and 1 if x >> bound. The parameter, slope, determines the range over
which the return value changes. The function used is the logistic function, where the return
value is 0.5 when x == bound and 0.95 when x95 == bound + log(19)/slope
}
\keyword{internal}
