% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcNLL_OffMoveMod}
\alias{CalcNLL_OffMoveMod}
\title{Calculate negative log-likelihood associated with length-at-age data, given
parameters for size-related movement growth model}
\usage{
CalcNLL_OffMoveMod(params)
}
\arguments{
\item{params}{}
}
\value{
NLL
}
\description{
This function negative log-likelihood associated with length-at-age data, given
parameters for size-related movement growth model (as described by Hesp et al., 2004)
}
\keyword{internal}
