% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{CalcPropMatureAtLength}
\alias{CalcPropMatureAtLength}
\title{Calculate proportions of mature fish in successive length classes}
\usage{
CalcPropMatureAtLength(MaxLen, LenInc, ObsLen, ObsMatCat)
}
\arguments{
\item{MaxLen}{maximum fish length}

\item{LenInc}{length class size increment}

\item{ObsLen}{vector of observed lengths}

\item{ObsMatCat}{vector of observed maturity categories (0=immature, 1=mature)}
}
\value{
length class lower bounds (lbnd), length class mid points (midpt),
proportions of mature fish in successive length classes (PropMat), lengths
for plotting maturity curve (plotlengths), length class sample sizes (LenClSampSize)
}
\description{
Calculates proportions of mature fish in successive length classes,
and outputs some variables useful for plotting
}
\examples{
# generate some synthetic maturity data
set.seed(123)
MaxAge = 20
MinLen = 0
MaxLen = 400
LenInc = 20
# single sex/combined sexes
nSexes = 1
nSamples = 300
CurveType = 1 # 1 = symmetric logistic, 2 = asymmetric logistic
Linf = 300
vbK = 0.3
tzero = 0
CVSizeAtAge = 0.1
GrowthParams = c(Linf, vbK, tzero, CVSizeAtAge)
L50 = 220
L95 = 270
Pmax = 1.0
MaturityParams = c(L50, L95, Pmax)
res=SimulateLengthAtMaturityData(nSamples, CurveType, nSexes, MaxAge, MinLen, MaxLen, LenInc, GrowthParams, MaturityParams)
ObsLen=res$ObsLen
ObsMatCat=res$ObsMatCat
CalcPropMatureAtLength(MaxLen, LenInc, ObsLen, ObsMatCat)
}
