% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{FitSeasonalGrowthModel}
\alias{FitSeasonalGrowthModel}
\title{Fit a seasonal growth curve to a sample of fish length-at-age data}
\usage{
FitSeasonalGrowthModel(
  params,
  nSexes,
  DataType,
  ObsAge,
  ObsLen,
  ObsMeanLen,
  ObsMeanLensd
)
}
\arguments{
\item{Linf, }{vbK, tzero, tc, C, Mod_sd (Numbers).}
}
\value{
nlmb (stored output from internal R nlminb optimisation function)
}
\description{
This function fits a Somers (1988) growth curve to a sample of fish length-at-age data
by minimising the negative log-likelihood associated with the parameters and data, using nlminb
}
\keyword{internal}
