% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{GetConfidenceLimitsForGrowthCurve}
\alias{GetConfidenceLimitsForGrowthCurve}
\title{Get confidence limits for the fitted growth curve.}
\usage{
GetConfidenceLimitsForGrowthCurve(
  GrowthEqn,
  nSexes,
  Ref_ages,
  params,
  vcov.params,
  ObsAge,
  ObsLen,
  ObsSex,
  plotages
)
}
\arguments{
\item{GrowthEqn}{1=von Bertalanffy, 2=Schnute, 3=Somers seasonal curve, 4=divergent growth model}

\item{nSexes}{number of sexes}

\item{Ref_ages}{Schnute growth curve reference ages (t1 and t2)}

\item{params}{estimated parameters}

\item{vcov.params}{estimated variance-covariance matrix}

\item{ObsAge}{observed individual fish age data (used for Growth eqn option 4)}

\item{ObsLen}{observed individual fish length data (used for Growth eqn option 4)}

\item{ObsSex}{observed individual fish sex (or area, inputted as sex) data (used for Growth eqn option 4)}

\item{plotages}{ages specified to produce estimates of uncertainty for estimated lengths}
}
\value{
scatter plot of length-at-age data
}
\description{
Use parametric resampling to get confidence limits for the fitted growth curve
}
