% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{GetOffMoveGrowthModResults}
\alias{GetOffMoveGrowthModResults}
\title{Fits growth model allowing for size-related movements of fish between 2 habitats using RTMB}
\usage{
GetOffMoveGrowthModResults(dat, params)
}
\arguments{
\item{dat}{list of all data inputs including following, eps (used in posfun_OffMoveMod penalty), lnL50MoveUp, lnL50MoveLw, lnSlopeMoveUp, lnSlopeMoveLw,
lnLinfUp, lnLinfLw, lnvbKUp, lnvbKLw, tzeroLw, tzeroUp, CVLw, CVUp (parameter bounds), DoJitter (boolean flag for jitter analysis), nJitters,
moveL50sd_jitter, moveSlopesd_jitter, Linfsd_jitter, vbKsd_jitter, tzerosd_jitter, CVsd_jitter, SampleSize, Habitat, ObsAges, ObsLengths,
nodes, weights}

\item{params}{lnmoveL50, lnmoveSlope, lnLinf, lnvbK, tzero, CV}
}
\value{
par, convergence, NLL, sdr, param_est, params_sd, EstParams, JitterRes (SummaryRes), RTMBReport
}
\description{
This function is for a von Bertalanffy growth model allowing for size-related movements of fish between 2 habitats,
(see Hesp et al., 2004). In addition to movement parameters, the model can correct for bias in parameter estimates
associated with non-representative sampling across the two habitats (i.e. one under sampled relative to the other),
assuming that sampling is representative within each habitat.
}
\examples{
library(RTMB)
library(L3Assess)
# *****************************************
# simulate some standard length-at-age data
# *****************************************
set.seed(123)
SampleSize=1000 # sample size for retained catches (and same number for released fish, if an MLL is specified)
MaxAge = 20
TimeStep = 1/12 # model timestep (e.g. 1 = annual, 1/12 = monthly)
NatMort = 4.22/12
FishMort = 0.1
MaxLen = 600
LenInc = 2
MLL=NA # (minimum legal length) # retention set to 1 for all lengths if MLL set to NA and retention parameters not specified
SelectivityType=2 # 1=selectivity inputted as vector, 2=asymptotic logistic selectivity curve
SelectivityAtLen = NA # selectivity vector
SelParams = c(100, 20) # L50, L95-L50 for gear selectivity
RetenParams = c(NA, NA) # L50, L95-L50 for retention
DiscMort = 0 # proportion of fish that die due to natural mortality
GrowthCurveType = 1 # 1 = von Bertalanffy, 2 = Schnute
Linf = 300
vbK = 0.5
CVSizeAtAge = 0.03
GrowthParams = c(Linf, vbK)
RefnceAges = NA
Res=SimLenAndAgeFreqData_EqMod(SampleSize, MaxAge, TimeStep, NatMort, FishMort, MaxLen, LenInc, MLL, SelectivityType,
                               SelParams, RetenParams, SelectivityAtLen, DiscMort, GrowthCurveType, GrowthParams, RefnceAges, CVSizeAtAge)
ObsAges = Res$ObsDecAgeRetCatch
ObsLengths = Res$ObsRandLenRetCatch
plot(ObsAges, ObsLengths, xlim=c(0,20), ylim=c(0,400))
# ***********************************************
# Assign each fish to inshore or offshore habitat
# ***********************************************
moveL50 = 230
moveSlope = 0.05
moveL95 = (log(19)/moveSlope) + moveL50
Habitat <- rep(0,SampleSize)
for (i in 1:SampleSize) {
  randnum = runif(1,0,1)
  tempProbOff = 1 / (1 + exp(-log(19) * (ObsLengths[i] - moveL50) / (moveL95 - moveL50)))
  if (randnum<tempProbOff) {
    Habitat[i] = 1
  } else {
    Habitat[i] = 0
  }
}
# ***************************************************************
# Assign relative sampling intensity in inshore and offshore area
# ***************************************************************
OffSampIntensity = 0.1
InshSampIntensity = 1
x = which(Habitat==0)
y = which(Habitat==1)
RandAgeInsh = ObsAges[x]
RandAgeOff = ObsAges[y]
RandLenInsh = ObsLengths[x]
RandLenOff = ObsLengths[y]
randnum=runif(length(x),0,1)
xx=which(randnum<InshSampIntensity)
randnum=runif(length(y),0,1)
yy=which(randnum<OffSampIntensity)
RandAgeInsh = RandAgeInsh[xx]
RandLenInsh = RandLenInsh[xx]
RandAgeOff = RandAgeOff[yy]
RandLenOff = RandLenOff[yy]
# plot random data
par(mfrow=c(2,2),mar=c(4,3,2,2))
plot(RandAgeInsh,RandLenInsh, xlim=c(0,20), ylim=c(0,400), col="red")
plot(RandAgeOff,RandLenOff, xlim=c(0,20), ylim=c(0,400), col="blue")
plot(RandAgeOff,RandLenOff, xlim=c(0,20), ylim=c(0,400), col="blue")
points(RandAgeInsh,RandLenInsh, col="red")
plot(RandAgeInsh,RandLenInsh, xlim=c(0,20), ylim=c(0,400), col="red")
points(RandAgeOff,RandLenOff, col="blue")
# Inshore
lbnd=Res$lbnd
InshLenCat = trunc(RandLenInsh/10)*10
histdat = hist(InshLenCat, breaks=c(lbnd,800), plot=F)
InshLenFreq = histdat$counts
breaks = histdat$breaks
plot(lbnd,InshLenFreq,"l")
# Offshore
OffLenCat = trunc(RandLenOff/10)*10
histdat = hist(OffLenCat, breaks=c(lbnd,800), plot=F)
OffLenFreq = histdat$counts
breaks = histdat$breaks
lines(lbnd,OffLenFreq,"l",col="blue")
ObsAge = c(RandAgeInsh,RandAgeOff)
ObsLen = c(RandLenInsh,RandLenOff)
Habitat = c(rep(0,length(RandLenInsh)),rep(1,length(RandLenOff)))
# *******************************************************************
# plot lengths at specified ages. Lengths of fish in offshore habitat
# expected to be larger at younger ages
# *******************************************************************
lbnd=Res$lbnd
par(mfrow=c(3,2))
for (SpecAge in 1:5) {
  x=which(RandAgeInsh >= SpecAge & RandAgeInsh < SpecAge+1)
  InshLenCat = trunc(RandLenInsh[x]/10)*10
  histdat = hist(InshLenCat, breaks=c(lbnd,800), plot=F)
  InshLenFreq = histdat$counts
  breaks = histdat$breaks
  x=which(RandAgeOff >= SpecAge & RandAgeOff < SpecAge+1)
  OffLenCat = trunc(RandLenOff[x]/10)*10
  histdat = hist(OffLenCat, breaks=c(lbnd,800), plot=F)
  OffLenFreq = histdat$counts
  breaks = histdat$breaks
  ymax=1.2*max(c(InshLenFreq, OffLenFreq))
  plot(lbnd,InshLenFreq,"l",ylim=c(0,ymax), main=paste("Age =",SpecAge))
  lines(lbnd,OffLenFreq, col="blue")
}
# ********************************************
# Fit traditional von Bertalanffy growth model
# ********************************************
nSexes = 1
DataType=1
params = c(log(400),log(0.3),0) # log(Linf), log(k), tzero
ObsAge = c(RandAgeInsh,RandAgeOff)
ObsLen = c(RandLenInsh,RandLenOff)
Habitat = c(rep(0,length(RandLenInsh)),rep(1,length(RandLenOff)))
TradvbRes=GetvonBertalanffyGrowthResults(params, nSexes, DataType, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLense=NA)
TradvbRes$ParamEst
# ***************************
# Fit offshore movement model
# ***************************
# specify data inputs
dat <- list(eps = 0.001,
            lnL50MoveUp = log(300),
            lnL50MoveLw = log(170),
            lnSlopeMoveUp = log(0.5),
            lnSlopeMoveLw = log(0.01),
            lnLinfUp = log(500),
            lnLinfLw = log(200),
            lnvbKUp = log(0.8),
            lnvbKLw = log(0.2),
            tzeroLw = -5,
            tzeroUp = 1,
            CVLw = 0.01,
            CVUp = 0.15,
            DoJitter = FALSE,
            nJitters = NA,
            moveL50sd_jitter = 20,
            moveSlopesd_jitter = 0.002,
            Linfsd_jitter = 20,
            vbKsd_jitter = 0.05,
            tzerosd_jitter = 0.05,
            CVsd_jitter = 0.005,
            SampleSize = length(Habitat),
            Habitat = Habitat,
            ObsAges = ObsAge,
            ObsLengths = ObsLen)
# specify parameters
params <- list(lnmoveL50 = log(200),
               lnmoveSlope = log(0.05),
               lnLinf = log(260),
               lnvbK = log(0.5),
               tzero = 0.0,
               CV = 0.05)
# Run RTMB model
FittedRes= GetOffMoveGrowthModResults(dat, params)
FittedRes$SummaryRes$JitterRes
FittedRes$SummaryRes$NLL
FittedRes$SummaryRes$EstParams
# plot expected mean lengths at age for each habitat
par(mfrow=c(2,2),mar=c(5,4,2,2))
# first habitat
x=which(dat$Habitat==0)
plot(dat$ObsAges[x],dat$ObsLengths[x], xlim=c(0,20), ylim=c(0,400),
     col="red",xlab="Age",ylab="Length",bty='n', cex.main=0.8, main="Juv Habitat")
points(dat$ObsAges[x], FittedRes$RTMBReport$ExpLen[x], col="black")
# second habitat
y=which(dat$Habitat==1)
plot(dat$ObsAges[y],dat$ObsLengths[y], xlim=c(0,20), ylim=c(0,400),
     col="blue",xlab="Age",ylab="Length",bty='n', cex.main=0.8, main="Adult Habitat")
points(dat$ObsAges[y], FittedRes$RTMBReport$ExpLen[y], col="black")
# curves, both habitats
plot(dat$ObsAges[x], FittedRes$RTMBReport$ExpLen[x], xlim=c(0,20), ylim=c(0,400),
     xlab="Age",ylab="Length",bty='n', col="red")
points(dat$ObsAges[y], FittedRes$RTMBReport$ExpLen[y], col="blue")
#  overall growth curves trad vs move model
# trad
Linf1 = FittedRes$SummaryRes$EstParams[3,1]
vbK1 = FittedRes$SummaryRes$EstParams[4,1]
tzero1 = FittedRes$SummaryRes$EstParams[5,1]
Ages = seq(0,20,0.1)
ExpLen1 = Linf1 * (1-exp(-vbK1 * (Ages-tzero1)))
plot(Ages, ExpLen1, xlim=c(0,20), ylim=c(0,300),
     xlab="Age",ylab="Length",bty='n', col="green")
# move mod
Linf2 = TradvbRes$ParamEst[1,1]
vbk2 =  TradvbRes$ParamEst[2,1]
tzero2 =  TradvbRes$ParamEst[3,1]
ExpLen2 = Linf2 * (1-exp(-vbk2 * (Ages-tzero2)))
points(Ages, ExpLen2, col="brown")
}
