% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{GetSchnuteGrowthResults}
\alias{GetSchnuteGrowthResults}
\title{Get outputs from a fitted Schnute growth curve.}
\usage{
GetSchnuteGrowthResults(
  params,
  nSexes,
  DataType,
  t1,
  t2,
  ObsAge,
  ObsLen,
  ObsMeanLen,
  ObsMeanLense
)
}
\arguments{
\item{params}{c(log(y1),log(y2),a,b) Schnute parameters}

\item{nSexes}{number of sexes}

\item{DataType}{1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis}

\item{t1}{first reference age}

\item{t2}{second reference age}

\item{ObsAge}{observed ages}

\item{ObsLen}{observed lengths}

\item{ObsMeanLen}{mean lengths for specified ages (i.e. as estimated from mixture analysis)}

\item{ObsMeanLense}{se for mean estimated mean lengths for specified ages (i.e. as estimated from mixture analysis)}
}
\value{
negative log-likelihood (nll), nlminb convergence diagnostic (convergence),
sample size (SampleSize), growth parameter estimates with lower and upper 95 percent confidence limits
(ParamEst), point estimates for growth parameters (params), variance-covariance matrix (vcov.params)
}
\description{
This function fits a Schnute growth curve to a sample of fish length-at-age data
by minimising the negative log-likelihood associated with the parameters and data,
using nlminb. It provides various statistical outputs in include convergence statistics,
parameter estimated and associated 95 percent confidence limits and associated variance-covariance matrix,
calculated using the MASS package.
}
\examples{
# Schnute growth equation - single sex length-at-age data
# simulate data (ignoring mortality and selectivity effects)
set.seed(123)
GrowthEqn = 2 # Schnute growth equation
nSexes = 1 # single or combined sex
nSamples = 500
MinAge = 0
MaxAge = 20
AgeStep = 1
Ref_ages = c(0.5,15)
Growth_params = c(50,400,0.2,0.5)
Growth_cv = 0.1
Res=SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# fit growth model and plot
ObsAge=Res$ObsAge
ObsLen=Res$ObsLen
par(mfrow=c(1,1))
plot(ObsAge,ObsLen)
DataType=1  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = log(c(80,400,0.2,0.5))
t1=0.5
t2=15
FittedRes=GetSchnuteGrowthResults(params, nSexes, DataType, t1, t2, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLense=NA)
FittedRes$ParamEst
# Schnute growth equation - single sex mean length at age data
# simulate data (ignoring mortality and selectivity effects)
DataType=2  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
nSamples = 2000
MinAge = 0
MaxAge = 3
AgeStep = 0.1
Ref_ages = c(0.1,2.5)
Growth_params = c(20,200,1.0,0.5)
Growth_cv = 0.1
Res=SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# fit growth model and plot
ObsAge=Res$ObsAge
ObsMeanLen=Res$ObsMeanLen
ObsMeanLense=Res$ObsMeanLense
plot(Res$ObsAgeCl,ObsMeanLen)
params = log(c(25,180,0.9,0.5))
t1=0.1
t2=2.5
ObsAge=Res$ObsAgeCl
ObsMeanLen=Res$ObsMeanLen
ObsMeanLense=Res$ObsMeanLense
FittedRes=GetSchnuteGrowthResults(params, nSexes, DataType, t1, t2, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLense)
FittedRes$ParamEst
# Schnute growth equation - 2 sexes, length-at-age data
# simulate data (ignoring mortality and selectivity effects)
set.seed(123)
GrowthEqn = 2 # Schnute equation
nSexes = 2 # separate sexes
nSamples = c(500,500)
MinAge = 1
MaxAge = 25
AgeStep = 1
Ref_ages = c(1,15)
Growth_params = c(100,120,500,550,0.2,0.2,0.3,0.4)
Growth_cv = c(0.1,0.1)
Res=SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# Fit Schnute growth curve to length at age dat
ObsAge=as.matrix(Res$ObsAge)
ObsLen=as.matrix(Res$ObsLen)
par(mfrow=c(1,1))
plot(ObsAge[1,],ObsLen[1,])
points(ObsAge[2,],ObsLen[2,],col="blue")
DataType=1  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = log(c(90,110,420,430,0.3,0.3,0.6,0.6))
t1=1
t2=15
FittedRes=GetSchnuteGrowthResults(params, nSexes, DataType, t1, t2, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLense=NA)
FittedRes$ParamEst
# Fit Schnute growth curve - 2 sexes mean length at age data
# simulate data (ignoring mortality and selectivity effects)
set.seed(123)
nSamples = c(2000,2000)
DataType=2  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
MinAge = 0.1
MaxAge = 3
AgeStep = 0.1
ObsAge = runif(nSamples, MinAge, MaxAge)
Growth_params = c(20,20,180,200,1.0,1.0,0.5,0.5)
Ref_ages = c(0.1,2.5)
Growth_cv = c(0.1,0.1)
Res=SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
par(mfrow=c(1,1))
plot(Res$ObsAgeCl, Res$FemObsMeanLen)
points(Res$ObsAgeCl, Res$MalObsMeanLen,col="blue")
t1=0.1
t2=2.5
params = log(c(20,30,190,210,0.9,1.1,0.5,0.5))
ObsAge=Res$ObsAgeCl
FemObsMeanLen = Res$FemObsMeanLen
MalObsMeanLen = Res$MalObsMeanLen
ObsMeanLen = as.matrix(t(data.frame(FemObsMeanLen=FemObsMeanLen,MalObsMeanLen=MalObsMeanLen)))
FemObsMeanLense = Res$FemObsMeanLense
MalObsMeanLense = Res$MalObsMeanLense
ObsMeanLense = as.matrix(t(data.frame(FemObsMeanLense=FemObsMeanLense,MalObsMeanLense=MalObsMeanLense)))
FittedRes=GetSchnuteGrowthResults(params, nSexes, DataType, t1, t2, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLense)
FittedRes$ParamEst
}
