% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{GetTaggingGrowthModelResults}
\alias{GetTaggingGrowthModelResults}
\title{Get statistical outputs from a fitted tagging growth model.}
\usage{
GetTaggingGrowthModelResults(
  params,
  nstep,
  Obs_delta_t,
  Obs_Initlen,
  Obs_Finlen,
  MaxAge,
  nobs
)
}
\arguments{
\item{params}{log(c(L50_1, L95_1, L50_2, L95_2)) double logistic model, or
log(c(Gaussian_A, Gaussian_u, Gaussian_sd)) Gaussian function, or log(c(vb_Linf, vb_K)) von Bertalanffy, or
log(c(Gomp_Linf, Gomp_G)) Gompertz}

\item{nstep}{number of numerical integration steps (higher number increases accuracy but reduces program speed)}

\item{Obs_delta_t}{observed durations at liberty for individual animals}

\item{Obs_Initlen}{observed initial lengths for individual animals}

\item{Obs_Finlen}{observed final lengths for individual animals}

\item{MaxAge}{specified maximum age for species}

\item{nobs}{number of observations}
}
\value{
negative log-likelihood (nll), nlminb convergence diagnostic (convergence),
growth parameter estimates with lower and upper 95 percent confidence limits (ParamEst),
point estimates for growth parameters (params), variance-covariance matrix (vcov.params),
observed differences between final and initial lengths (Obs_delta_Length), observed final
lengths (Est_Final_Length), specified set of initial lengths for plotting model results
(plot_Init_Length), estimated annual growth given initial length, for plotting model results
(plot_Length_Inc), integer ages for plotting model results (plot_IntAge), estimated lengths
at integer ages for plotting (plot_EstLenAtIntAge)
}
\description{
This function fits a model to fish tag-recapture data
by minimising the negative log-likelihood associated with the parameters and data,
using nlminb. It provides various statistical outputs in include convergence statistics,
parameter estimated and associated 95 percent confidence limits and associated variance-covariance matrix,
calculated using the MASS package.
}
\examples{
set.seed(123)
nstep = 50 # number of steps for numerical integration
MaxLen = 240
Gaussian_A = 0.1
Gaussian_u = 80
Gaussian_sd = 40
StandDev = 10
GrowthCrvChoice = 2 # 1 = double logistic, 2 = Gaussian function, 3 = von Bertalanffy growth curve, 4 = Gompertz growth curve
params = log(c(Gaussian_A, Gaussian_u, Gaussian_sd, StandDev))
nobs = 200
CalculationStage = 1
res=SimulateTagRecaptureData(GrowthCrvChoice, nstep, nobs, MaxLen, params)
Obs_delta_t=res$Obs_delta_t
Obs_Initlen=res$Obs_Initlen
Obs_Finlen=res$Obs_Finlen
params = log(c(0.1, 80, 40, 10))
MaxAge = 20
GetTaggingGrowthModelResults(params, nstep, Obs_delta_t, Obs_Initlen, Obs_Finlen, MaxAge, nobs)
}
