% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{LogisticEqnAgeAtMaturity2}
\alias{LogisticEqnAgeAtMaturity2}
\title{Logistic curve for age at maturity (for plotting)}
\usage{
LogisticEqnAgeAtMaturity2(params, nSexes, plotages)
}
\arguments{
\item{params}{c(A50, A95) ages at 50 and 95 percent maturity}

\item{nSexes}{number of sexes}

\item{plotages}{age classes for plotting}
}
\value{
probability of maturity for observed ages. Used in calculation of objective function
for length at maturity model
}
\description{
This function applies a logistic curve for describing probability of maturity at age, used for plotting
maturity results
}
\keyword{internal}
