% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{LogisticEqnLengthAtMaturity}
\alias{LogisticEqnLengthAtMaturity}
\title{Logistic curve for length at maturity}
\usage{
LogisticEqnLengthAtMaturity(params)
}
\arguments{
\item{params}{c(L50, L95) lengths at 50 and 95 percent maturity}

\item{nSexes}{number of sexes}
}
\value{
probability of maturity for observed lengths. Used in calculation of objective function
for length at maturity model
}
\description{
This function applies a logistic curve for describing probability of maturity at length
}
\keyword{internal}
