% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{LogisticEqnLengthAtMaturity2}
\alias{LogisticEqnLengthAtMaturity2}
\title{Logistic curve for length at maturity (for plotting)}
\usage{
LogisticEqnLengthAtMaturity2(params, CurveType, nSexes, plotlengthrange)
}
\arguments{
\item{params}{c(L50, L95) or c(Pmax, L50, L95) for CurveType=1, or c(Q, B, V) or c(Pmax, Q, B, V) for CurveType=2}

\item{CurveType}{1 = symmetric, 2 = asymmetric}

\item{nSexes}{number of sexes}

\item{plotlengthrange}{lower and upper length class range for plotting}
}
\value{
probability of maturity for specified lengths for plotting, used for plotting maturity results
for length at maturity model
}
\description{
This function applies a logistic curve for describing probability of maturity at length,
used for plotting
}
\keyword{internal}
