% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{PlotFittedGrowthCurve}
\alias{PlotFittedGrowthCurve}
\title{Plot fitted growth curve to fish length-at-age data.}
\usage{
PlotFittedGrowthCurve(
  DataType,
  nSexes,
  GrowthEqn,
  ObsAge,
  ObsLen,
  ObsSex,
  ObsMeanLen,
  ObsMeanLense,
  params,
  Ref_ages,
  plotages,
  ymax,
  xmax,
  yint,
  xint,
  GraphTitle,
  xaxis_lab,
  yaxis_lab,
  PlotCLs,
  FittedRes
)
}
\arguments{
\item{DataType}{# 1=lengths at age data for individual fish (single sex, or sexes recorded),
2=mean length at age and sd data from mixture analysis, 3=lengths at age data for individual fish
(two sex, sexes not recorded)}

\item{nSexes}{number of sexes}

\item{GrowthEqn}{1=von Bertalanffy, 2=Schnute, 3=Somers seasonal growth curve, 4=Divergent growth curve}

\item{ObsAge}{observed individual fish ages}

\item{ObsLen}{observed individual fish sex (required for GrowthEqn=4), otherwise set to NA}

\item{ObsMeanLen}{mean lengths for specified ages (i.e. as estimated from mixture analysis)}

\item{params}{estimated growth parameters}

\item{plotages}{specified ages for plotting}

\item{ymax}{maximum value for y axis}

\item{xmax}{maximum value for x axis}

\item{yint}{interval for y axis}

\item{xint}{interval for x axis}

\item{GraphTitle}{graph title}

\item{xaxis_lab}{label for x axis}

\item{yaxis_lab}{label for y axis}

\item{PlotCLs}{logical True=plot 95 percent confidence limits for curve}

\item{ObsMeanLensd}{sd for mean estimated mean lengths for specified ages (i.e. as estimated from mixture analysis)}
}
\value{
fitted curve on scatter plot with length-at-age data
}
\description{
Plot fitted growth curve to fish length-at-age data.
}
\examples{
# von Bertalanffy growth equation
# simulate data (ignoring mortality and selectivity effects)
GrowthEqn = 1 # von Bertalanffy growth equation
nSexes = 1 # single or combined sex
nSamples = 300
MinAge = 1
MaxAge = 20
AgeStep = 1
Linf = 300
vbK = 0.3
tzero = 0
Growth_params = c(Linf, vbK, tzero)
Ref_ages = NA
Growth_cv = 0.08
Res = SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# fit growth model and plot
DataType=1  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = c(log(400),log(0.3),0) # log(Linf), log(k), tzero
ObsAge = Res$ObsAge
ObsLen = Res$ObsLen

FittedRes=GetvonBertalanffyGrowthResults(params, nSexes, DataType, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLense=NA)
plotages=seq(0, MaxAge,0.1)
par(mfrow=c(1,1))
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen=NA,
                      ObsMeanLense=NA, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
DataType=2  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = c(log(400),log(0.3),0) # log(Linf), log(k), tzero
ObsAge = Res$ObsAgeCl
ObsMeanLen = Res$ObsMeanLen
ObsMeanLense = Res$ObsMeanLense
FittedRes=GetvonBertalanffyGrowthResults(params, nSexes, DataType, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLense)
plotages=seq(0, MaxAge,0.1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen,
                      ObsMeanLense, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
nSexes = 2 # Separate sexes
nSamples = c(500,500) # if different, order so NAs are last for smaller sample
MinAge = 1
MaxAge = 20
AgeStep = 1
Linf = c(300,250)
vbK = c(0.3,0.3)
tzero = c(0,0)
GrowthEqn = 1
Growth_params = c(Linf, vbK, tzero)
Ref_ages = NA
Growth_cv = c(0.08,0.08)
Res = SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# fit growth model and plot
DataType=1  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
ObsAge = as.matrix(Res$ObsAge)
ObsLen = as.matrix(Res$ObsLen)
params = c(log(c(300,300)),log(c(0.3,0.3)),c(0,0)) # log(Linf), log(k), tzero
FittedRes=GetvonBertalanffyGrowthResults(params, nSexes, DataType, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLense=NA)
plotages=seq(0, MaxAge,0.1)
par(mfrow=c(1,2))
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen,
                      ObsMeanLense, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
DataType=2  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = c(log(c(300,300)),log(c(0.3,0.3)),c(0,0)) # log(Linf), log(k), tzero
ObsAge = Res$ObsAgeCl
FemObsMeanLen = Res$FemObsMeanLen
MalObsMeanLen = Res$MalObsMeanLen
ObsMeanLen = as.matrix(t(data.frame(FemObsMeanLen=FemObsMeanLen,MalObsMeanLen=MalObsMeanLen)))
FemObsMeanLense = Res$FemObsMeanLense
MalObsMeanLense = Res$MalObsMeanLense
ObsMeanLense = as.matrix(t(data.frame(FemObsMeanLense=FemObsMeanLense,MalObsMeanLense=MalObsMeanLense)))
FittedRes=GetvonBertalanffyGrowthResults(params, nSexes, DataType, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLense)
plotages=seq(0, MaxAge,0.1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen,
                      ObsMeanLense, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
# Schnute growth equation - single sex length-at-age data
# simulate data (ignoring mortality and selectivity effects)
set.seed(123)
GrowthEqn = 2 # Schnute growth equation
nSexes = 1 # single or combined sex
nSamples = 500
MinAge = 0
MaxAge = 20
AgeStep = 1
Ref_ages = c(0.5,15)
Growth_params = c(50,400,0.2,0.5)
Growth_cv = 0.1
Res=SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# fit growth model and plot
ObsAge=Res$ObsAge
ObsLen=Res$ObsLen
par(mfrow=c(1,1))
plot(ObsAge,ObsLen)
DataType=1  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = log(c(80,400,0.2,0.5))
t1=0.5
t2=15
FittedRes=GetSchnuteGrowthResults(params, nSexes, DataType, t1, t2, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLense=NA)
FittedRes$ParamEst
FittedRes$cor.params
plotages=seq(0, MaxAge,0.1)
par(mfrow=c(1,1))
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen=NA,
                      ObsMeanLense=NA, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
# Schnute growth equation - single sex mean length at age data
# simulate data (ignoring mortality and selectivity effects)
DataType=2  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
nSamples = 2000
MinAge = 0
MaxAge = 3
AgeStep = 0.1
Ref_ages = c(0.1,2.5)
Growth_params = c(20,200,1.0,0.5)
Growth_cv = 0.1
Res=SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# fit growth model and plot
ObsAge=Res$ObsAge
ObsMeanLen=Res$ObsMeanLen
ObsMeanLense=Res$ObsMeanLense
plot(Res$ObsAgeCl,ObsMeanLen)
params = log(c(25,180,0.9,0.5))
t1=0.1
t2=2.5
ObsAge=Res$ObsAgeCl
ObsMeanLen=Res$ObsMeanLen
ObsMeanLense=Res$ObsMeanLense
FittedRes=GetSchnuteGrowthResults(params, nSexes, DataType, t1, t2, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLense)
FittedRes$ParamEst
FittedRes$cor.params
plotages=seq(0, MaxAge,0.1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen,
                      ObsMeanLense, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
# Schnute growth equation - 2 sexes, length-at-age data
# simulate data (ignoring mortality and selectivity effects)
set.seed(123)
GrowthEqn = 2 # Schnute equation
nSexes = 2 # separate sexes
nSamples = c(500,500)
MinAge = 1
MaxAge = 25
AgeStep = 1
Ref_ages = c(1,15)
Growth_params = c(100,120,500,550,0.2,0.2,0.3,0.4)
Growth_cv = c(0.1,0.1)
Res=SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# Fit Schnute growth curve to length at age dat
ObsAge=as.matrix(Res$ObsAge)
ObsLen=as.matrix(Res$ObsLen)
par(mfrow=c(1,1))
plot(ObsAge[1,],ObsLen[1,])
points(ObsAge[2,],ObsLen[2,],col="blue")
DataType=1  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = log(c(90,110,420,430,0.3,0.3,0.6,0.6))
t1=1
t2=15
FittedRes=GetSchnuteGrowthResults(params, nSexes, DataType, t1, t2, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLense=NA)
FittedRes$ParamEst
FittedRes$cor.params
par(mfrow=c(1,2))
plotages=seq(0, MaxAge,1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen=NA,
                      ObsMeanLense=NA, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
# Fit Schnute growth curve - 2 sexes mean length at age data
# simulate data (ignoring mortality and selectivity effects)
set.seed(123)
nSamples = c(2000,2000)
DataType=2  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
MinAge = 0.1
MaxAge = 3
AgeStep = 0.1
ObsAge = runif(nSamples, MinAge, MaxAge)
Growth_params = c(20,20,180,200,1.0,1.0,0.5,0.5)
Ref_ages = c(0.1,2.5)
Growth_cv = c(0.1,0.1)
Res=SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
par(mfrow=c(1,1))
plot(Res$ObsAgeCl, Res$FemObsMeanLen)
points(Res$ObsAgeCl, Res$MalObsMeanLen,col="blue")
t1=0.1
t2=2.5
params = log(c(20,30,190,210,0.9,1.1,0.5,0.5))
ObsAge=Res$ObsAgeCl
FemObsMeanLen = Res$FemObsMeanLen
MalObsMeanLen = Res$MalObsMeanLen
ObsMeanLen = as.matrix(t(data.frame(FemObsMeanLen=FemObsMeanLen,MalObsMeanLen=MalObsMeanLen)))
FemObsMeanLense = Res$FemObsMeanLense
MalObsMeanLense = Res$MalObsMeanLense
ObsMeanLense = as.matrix(t(data.frame(FemObsMeanLense=FemObsMeanLense,MalObsMeanLense=MalObsMeanLense)))
FittedRes=GetSchnuteGrowthResults(params, nSexes, DataType, t1, t2, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLense)
FittedRes$ParamEst
FittedRes$cor.params
par(mfrow=c(1,2))
plotages=seq(0, MaxAge,0.1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen,
                      ObsMeanLense, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
# Seasonal growth curve
# simulate data (ignoring mortality and selectivity effects)
GrowthEqn=3 # Seasonal growth curve
nSexes=1 # single or combined sex
nSamples = 1000
MinAge = 0
MaxAge = 3
AgeStep = 1/12
Linf = 150
vbK = 1
tzero = 0
tc = 0.25
C = 0.8
Ref_ages=NA
Growth_params=c(Linf,vbK,tzero,tc,C)
Growth_cv = 0.1
Res=SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# fit growth model and plot
DataType = 1
nSexes=1
ObsAge = Res$ObsAge
ObsLen = Res$ObsLen
params = c(log(160),log(0.8),0,0.25,0.8) #Linf, vbK, tzero, tc, C
# fit seasonal growth curve to mean length at age data
FittedRes=GetSeasonalGrowthResults(params, nSexes, DataType, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLense=NA)
par(mfrow=c(1,1))
plotages=seq(0, MaxAge,0.1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen=NA, ObsMeanLense=NA,
                      params, Ref_ages, plotages, ymax=200, xmax=NA, yint=50, xint=NA, GraphTitle=NA,
                      xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
DataType=2  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = c(log(160),log(0.8),0,0.25,0.8) #Linf, vbK, tzero, tc, C
ObsAge=Res$ObsAgeCl
ObsMeanLen=Res$ObsMeanLen
ObsMeanLense=Res$ObsMeanLense
FittedRes=GetSeasonalGrowthResults(params, nSexes, DataType, ObsAge, ObsLen, ObsMeanLen, ObsMeanLense)
plotages=seq(0, MaxAge,0.1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen,
                      ObsMeanLense, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
GrowthEqn=3 # Seasonal growth curve
nSexes=2 # separate sexes
nSamples = c(500,500)
MinAge = 0
MaxAge = 3
AgeStep = 1/12
Linf = c(150,160)
vbK = c(1,1)
tzero = c(0,0)
tc = c(0.25,0.25)
C = c(0.8,0.8)
Ref_ages=NA
Growth_params=c(Linf,vbK,tzero,tc,C)
Growth_cv = c(0.1,0.1)
Res=SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# fit growth model and plot
DataType = 1
nSexes=2
ObsAge=as.matrix(Res$ObsAge)
ObsLen=as.matrix(Res$ObsLen)
params = c(log(c(160,150)),log(c(0.8,0.8)),c(0,0),c(0.25,0.25),c(0.8,0.8)) #Linf, vbK, tzero, tc, C
# fit seasonal growth curve to mean length at age data
FittedRes=GetSeasonalGrowthResults(params, nSexes, DataType, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLense=NA)
par(mfrow=c(1,2))
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen=NA,
                      ObsMeanLense=NA, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
DataType=2  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
params = c(log(c(160,150)),log(c(0.8,0.8)),c(0,0),c(0.25,0.25),c(0.8,0.8)) #Linf, vbK, tzero, tc, C
ObsAge=Res$ObsAgeCl
FemObsMeanLen = Res$FemObsMeanLen
MalObsMeanLen = Res$MalObsMeanLen
ObsMeanLen = as.matrix(t(data.frame(FemObsMeanLen=FemObsMeanLen,MalObsMeanLen=MalObsMeanLen)))
FemObsMeanLense = Res$FemObsMeanLense
MalObsMeanLense = Res$MalObsMeanLense
ObsMeanLense = as.matrix(t(data.frame(FemObsMeanLense=FemObsMeanLense,MalObsMeanLense=MalObsMeanLense)))
FittedRes=GetSeasonalGrowthResults(params, nSexes, DataType, ObsAge, ObsLen=NA, ObsMeanLen, ObsMeanLense)
plotages=seq(0, MaxAge,0.1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen,
                      ObsMeanLense, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
# simulate data allowing for divergent growth
set.seed(123)
GrowthEqn = 4 # von Bertalanffy growth equation with divergent growth
nSexes = 2 # single or combined sex
nSamples = 500
MinAge = 0
MaxAge = 20
AgeStep = 1
Linf = c(1000,500)
vbK = c(0.15,0.15)
tzero1 = 0
tdiverge = 1
Growth_params = c(Linf, vbK, tzero1, tdiverge)
Ref_ages = NA
Growth_cv = 0.1
Res = SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
x=which(Res$ObsDat$ObsSex==1)
plot(Res$ObsAge[x], Res$ObsLen[x])
y=which(Res$ObsDat$ObsSex==2)
points(Res$ObsAge[y], Res$ObsLen[y],col="blue")
# fit growth model
InitLinf = c(1050,500)
# InitvbK = c(0.13,0.13) # separate vbK option
InitvbK = 0.13 # single vbK option
Inittzero1 = 0
Inittdiverge = 4
params = c(log(InitLinf), log(InitvbK), Inittzero1, log(Inittdiverge))
ObsAge = Res$ObsDat$ObsAge
ObsLen = Res$ObsDat$ObsLen
ObsSex = Res$ObsDat$ObsSex
FittedRes=GetDivergeGrowthModelResults(params, ObsAge, ObsLen, ObsSex)
FittedRes$Boot_ParamEst
DataType=1
plotages = seq(MinAge,MaxAge,0.1)
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex, ObsMeanLen,
                      ObsMeanLense, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
# Simulate data for separate sexes, fit separate von Bertalanffy growth curves, where sex is not specified in data
# 2 sexes
GrowthEqn = 1
nSamples = c(300,300)
nSexes = 2
MinAge = 1
MaxAge = 20
AgeStep = 1
Linf = c(400,600)
vbK = c(0.3,0.3)
tzero = c(0,0)
Growth_params = c(Linf,vbK,tzero)
Growth_cv = c(0.08,0.08)
Res = SimulateLengthAtAgeData(GrowthEqn, nSamples, nSexes, MinAge, MaxAge, AgeStep, Ref_ages, Growth_params, Growth_cv)
# plot lengths at age
for (i in 1:nSexes) {
  ObsAge=as.vector(unlist(Res$ObsAge[i,]))
  ObsLen=as.vector(unlist(Res$ObsLen[i,]))
  if (i==1) {
    plot(ObsAge, ObsLen, ylim=c(0,800))
  } else {
    points(ObsAge, ObsLen, col="blue")
  }
}
ObsAge=c(as.vector(unlist(Res$ObsAge[1,])),as.vector(unlist(Res$ObsAge[2,])))
ObsLen=c(as.vector(unlist(Res$ObsLen[1,])),as.vector(unlist(Res$ObsLen[2,])))
DataType=1  # 1=lengths at age data for individual fish, 2=mean length at age and sd data from mixture analysis
nSexes = 2 # separate sexes, but sex unknown
# note, model has 7, not 6 params, with estimated sd, to distinguish this from standard model
params = c(log(c(250,800)),log(c(0.3,0.3)),c(0,0), log(20)) # log(Linf), log(k), tzero, common sd
FittedRes=GetvonBertalanffyGrowthResults(params, nSexes, DataType, ObsAge, ObsLen, ObsMeanLen=NA, ObsMeanLense=NA)
plotages=seq(0, MaxAge,0.1)
par(mfrow=c(1,1),mar=c(4,3,2,2))
PlotFittedGrowthCurve(DataType, nSexes, GrowthEqn, ObsAge, ObsLen, ObsSex=NA, ObsMeanLen,
                      ObsMeanLense, params, Ref_ages, plotages, ymax=NA, xmax=NA, yint=NA, xint=NA,
                      GraphTitle=NA, xaxis_lab=NA, yaxis_lab=NA, PlotCLs=T, FittedRes)
}
