% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{SimulateLengthAtMaturityData}
\alias{SimulateLengthAtMaturityData}
\title{Simulate length at maturity data}
\usage{
SimulateLengthAtMaturityData(
  nSamples,
  CurveType,
  nSexes,
  MaxAge,
  MinLen,
  MaxLen,
  LenInc,
  GrowthParams,
  MaturityParams
)
}
\arguments{
\item{nSamples}{number of required samples}

\item{CurveType}{1 = symmetric logistic, 2 = asymmetric logistic}

\item{nSexes}{number of sexes}

\item{MaxAge}{maximum age}

\item{MinLen}{minimum length}

\item{MaxLen}{maximum length}

\item{LenInc}{length class interval}

\item{GrowthParams}{von Bertalanffy growth parameters and cv for mean size at age}

\item{MaturityParams}{for one or both sexes, for symmetric logistic (L50, L95, Pmax), or asymmetric logistic (Q, B, V, Pmax)}
}
\value{
Random fish lengths (ObsLen), maturity category, 1 or 0 (ObsMatCat),
length class lower bound (lbnd), length class mid point (midpt), proportion mature in length classes,
(PropMat),  plotlengths (lengths to plot maturity curve), length class sample sizes (LenClSampSize)
}
\description{
Simulate length at maturity data, given specified growth, size at maturity parameters for required sample size. Can
simulate data for one or two sexes, and based on a symmetric or asymmetric logistic curve.
}
\examples{
# simulate maturity data (ignoring mortality and selectivity effects)
nSamples = 300
CurveType = 1 # 1 = symmetric logistic, 2 = asymmetric logistic
nSexes = 1 # single or combined sex
MaxAge = 20
MinLen = 0
MaxLen = 500
LenInc = 50
Linf = 300 # von Bertalanffy growth equation
vbK = 0.3
tzero = 0
Growth_cv = 0.08
GrowthParams = c(Linf, vbK, tzero, Growth_cv)
L50 = 180
L95 = 200
Pmax = 1.0
MaturityParams = c(L50,L95,Pmax)
Res = SimulateLengthAtMaturityData(nSamples, CurveType, nSexes, MaxAge, MinLen, MaxLen, LenInc, GrowthParams, MaturityParams)
# nSamples = 300
# CurveType = 2 # 1 = symmetric logistic, 2 = asymmetric logistic
# nSexes = 1 # single or combined sex
# MaxAge = 20
# MinLen = 0
# MaxLen = 500
# LenInc = 50
# Linf = 300 # von Bertalanffy growth equation
# vbK = 0.3
# tzero = 0
# Growth_cv = 0.08
# GrowthParams = c(Linf, vbK, tzero, Growth_cv)
# Q = 1.0
# B = 0.01
# V = 0.5
# Pmax = 1.0
# MaturityParams = c(Q, B, V, Pmax)
# Res = SimulateLengthAtMaturityData(nSamples, CurveType, nSexes, MaxAge, MinLen, MaxLen, LenInc, GrowthParams, MaturityParams)
# nSamples = c(300, 300)
# nSexes = 2 # Separate sexes
# CurveType = 1 # 1 = symmetric logistic, 2 = asymmetric logistic
# MaxAge = 20
# MinLen = 0
# MaxLen = 500
# LenInc = 50
# Linf = c(300, 350)
# vbK = c(0.3, 0.3)
# tzero = c(0, 0)
# Growth_cv = c(0.08, 0.08)
# GrowthParams = c(Linf, vbK, tzero, Growth_cv)
# L50 = c(180, 190)
# L95 = c(200, 210)
# Pmax = c(1.0, 1.0)
# MaturityParams = c(L50,L95,Pmax)
# Res = SimulateLengthAtMaturityData(nSamples, CurveType, nSexes, MaxAge, MinLen, MaxLen, LenInc, GrowthParams, MaturityParams)
}
