% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_biology_methods.R
\name{SimulateTagRecaptureData}
\alias{SimulateTagRecaptureData}
\title{Simulate some tag-recapture data}
\usage{
SimulateTagRecaptureData(GrowthCrvChoice, nstep, nobs, MaxLen, params)
}
\arguments{
\item{GrowthCrvChoice}{1=double logistic model, 2=Gaussian function, 3=von Bertalanffy, 4=Gompertz}

\item{nstep}{number of numerical integration steps (higher number increases accuracy but reduces program speed)}

\item{nobs}{number of observations}

\item{MaxLen}{maximum length}

\item{params}{log(c(L50_1, L95_1, L50_2, L95_2)) double logistic model, or
log(c(Gaussian_A, Gaussian_u, Gaussian_sd)) Gaussian function, or log(c(vb_Linf, vb_K)) von Bertalanffy, or
log(c(Gomp_Linf, Gomp_G)) Gompertz}
}
\value{
simulated tag-recapture data for individuals animals, including random a set of initial lengths (Obs_Initlen),
final lengths (Obs_Finlen) and durations at liberty (Obs_delta_t), and the 'true' relationships (for two
durations at liberty), using for generating the random tag-recapture data.
}
\description{
Simulate tag-recapture data with varying times at liberty,
to which a growth curve may be fitted
}
\examples{
set.seed(123)
nstep = 50 # number of steps for numerical integration
MaxLen = 240
Gaussian_A = 0.1
Gaussian_u = 80
Gaussian_sd = 40
StandDev = 10
GrowthCrvChoice = 2 # 1 = double logistic, 2 = Gaussian function, 3 = von Bertalanffy growth curve, 4 = Gompertz growth curve
params = log(c(Gaussian_A, Gaussian_u, Gaussian_sd, StandDev))
nobs = 200
CalculationStage = 1
res=SimulateTagRecaptureData(GrowthCrvChoice, nstep, nobs, MaxLen, params)
plot(res$Obs_Initlen, res$Obs_Finlen, pch=16, cex=0.6)
lines(res$Initlen_line, res$Exp_Finlen2, col="blue")
lines(res$Initlen_line2, res$Exp_Finlen3, col="blue")
}
